Module snes::addr

Info

Everything here will fetch and read as needed. These will change the State in a number of ways such as incrementing the State::pc, or modifying the State::dbr.

The SNES has a 24 bit address space but only 16 bit registers. Values are combined using a number of address modes, registers and indirection to access the full 24 bit range.

For now I have commented out the more complex address modes and left the simple modes implemented. Just so I can progress on other parts of the system. These will be worked on slowly as I find more information.

These are difficult to implement correctly. Clear documentation on these has been difficult to find. Handling things like indexing overflow will be dealth with over time.

Testing the complex modes needs a lot of set up and precision. I am slowly working on this, but it’s hard to know if the tests are correct.

Extra Notes

SNES addresses will be shown as bank : address where bank is the high byte, and address is the two low bytes.

Functions

absolute

Absolute

absolute_indexed_indirect

Todo

absolute_indexed_x

Todo

absolute_indexed_y

Todo

absolute_indirect

Todo

absolute_long

Todo

absolute_long_indexed

Todo

direct

Direct Page

direct_indexed_indirect

Todo

direct_indexed_x

Todo

direct_indexed_y

Todo

direct_indirect

Todo

direct_indirect_indexed

Todo

direct_indirect_long

Todo

immediate16

Immediate

pc_relative

Program counter relative

pc_relative_long

Todo

stack_relative

stack pointer relative