Module reactor_rt::assembly

Module containing the API to initialize a reactor program.

Macros

unsafe_iter_bank

Iterates a bank, produces an Iterator<Item=&mut Port<_>>. Does not explicitly borrow the bank, which is unsafe, but we trust the code generator to fail if a port is both on the LHS and RHS of a connection.

Structs

AssemblyCtx

Helper struct to assemble reactors during initialization. One assembly context is used per reactor, they can’t be shared.

AssemblyError

An error occurring during initialization of the reactor program. Should never occur unless the graph is built by hand, and not by a Lingua Franca compiler.

ComponentCreator

Creates the components of a reactor.

DependencyDeclarator

GlobalReactionId

Global identifier for a reaction.

TriggerId

The ID of a trigger component.

Traits

ReactorInitializer

Wrapper around the user struct for safe dispatch.

Type Definitions

AssemblyResult