Module rest::basis_io

Modules

basic_math 🔒

basis_list

This mod is designed to check if a given basis set name is available in basissetexchange.org. If the given basis set name is not a exact match, a similar legal name will be suggested. The basis set name is not case sensitive.

bse_downloader

This mod is designed to download basis set or auxiliary basis set from basissetexchange.org if required basis sets for elements are missing in local.

etb

Generate even-tempered auxiliary Gaussian basis (ETB).
This work references df module of PySCF.

Structs

BasCell

#BasCell BasCell has the same organization structure of the basis set files from the basis set exchange (BSE)

BasCellRaw

BasInfo

BasInfo

Basis4Elem

#Basis4Elem Basis4Elem contains all informations of the GTO basis functions located in a given element self.electron_shells: the GTO basis functions organized cell by cell BasCell self.references: The reference of each basis cell. self.global_index: The global index for the given atom

Basis4ElemRaw

RefCell

Functions

_gaussian_int

cartesian_gto_1st_batch

cartesian_gto_1st_batch_par

cartesian_gto_1st_batch_par_v02

cartesian_gto_1st_batch_v02

cartesian_gto_1st_batch_v03

cartesian_gto_1st_cint

cartesian_gto_1st_cint_batch

cartesian_gto_1st_cint_batch_serial

cartesian_gto_1st_cint_batch_serial_v03

cartesian_gto_1st_std

cartesian_gto_batch

cartesian_gto_batch_par

cartesian_gto_batch_par_v02

cartesian_gto_batch_par_v03

cartesian_gto_batch_v03

cartesian_gto_batch_v04

cartesian_gto_cint

cartesian_gto_cint_batch

cartesian_gto_cint_batch_rayon

cartesian_gto_cint_batch_v04

cartesian_gto_std

cint_norm_factor

Normalization factor for GTO radial part g=r^l e^{-\alpha r^2}

gto_1st_value

gto_value

gto_value_debug

spheric_gto_1st_value_batch

spheric_gto_1st_value_batch_serial

spheric_gto_1st_value_serial

spheric_gto_value_matrixfull

spheric_gto_value_matrixfull_serial

spheric_gto_value_serial