Module rest::basis_io::etb

Generate even-tempered auxiliary Gaussian basis (ETB).
This work references df module of PySCF.

Structs

InfoV2

This structure works to hold results generated by etb_gen_for_atom_list. It includes an hashmap named ‘elements’ where the keys are element names and the values are structures named crate::basis_io::Basis4Elem which stores basis set information for certain elements.

Functions

etb_gen_contracted

Generate even-tempered auxiliary Gaussian basis for a single electron shell.

etb_gen_for_atom_list

etb_gen_for_atom_list_contracted

Generate even-tempered auxiliary Gaussian basis (ETB) for the given list of atoms.

etb_gen_for_atom_list_primitive

Generate even-tempered auxiliary Gaussian basis (ETB) for the given list of atoms.

etb_gen_primitive

etbs_gen_contracted

Generate even-tempered auxiliary Gaussian basis for a single atom.

etbs_gen_primitive

Generate even-tempered auxiliary Gaussian basis for a single atom.

get_basis_from_mol

Get corresponding basis set of the given atom from crate::molecule_io::Molecule.

get_etb_elem

Get elements of which the ETBs to be generated according to the start atom. Returns a vector of elements.