Module rust_action_heroes::bundle::movement

MovementBundle is an overloaded Bundle which initializes a bunch of Channel-based systems

All of which require some amount of Channel interaction, but that's about the only unifying thing.

There are three channels, and it's not like all of the systems here require writing/reading all three of them.

What I'm trying to say is if you think this code should be organized better I also think that but I haven't gotten around to it.

Structs

MovementBundle
MovementBundle initializes all systems which require event channels.