Struct wasm_svg_graphics::color::Color

pub struct Color { /* fields omitted */ }

Represents colors within the WASM SVG GRAPHICS lib

Methods

impl Color

pub fn new(red: u8, green: u8, blue: u8) -> Color

Constructor for color

pub fn from_hex(input_string: &str) -> Option<Color>

Constructor for color from hex

Example

use wasm_svg_graphics::color::Color;

// 6 character definition
let black_coral = Color::from_hex("565676");

// 7 character definition
let pale_lavender = Color::from_hex("#d8dcff");

// Generated with https://coolors.co/

pub fn r(&self) -> u8

Returns the red component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.r()); // 172

pub fn red(&self) -> u8

Returns the red component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.red()); // 172

pub fn g(&self) -> u8

Returns the green component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.g()); // 159

pub fn green(&self) -> u8

Returns the green component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.green()); // 159

pub fn b(&self) -> u8

Returns the blue component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.b()); // 187

pub fn blue(&self) -> u8

Returns the blue component of the color

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.blue()); // 187

pub fn rgb(&self) -> (u8, u8, u8)

Returns a tuple with red, green and blue, respectively

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{:?}", pastel_purple.rgb()); // (172, 159, 187)

pub fn to_rgb_string(&self) -> String

Returns a string a string containing the u8 variant of the colors

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.to_rgb_string()); // rgb(172, 159, 187)

pub fn to_hex_string(&self) -> String

Returns a string a string containing the hex variant of the colors

Example

use wasm_svg_graphics::color::Color;

let pastel_purple = Color::new(172, 159, 187);

println!("{}", pastel_purple.to_hex_string()); // #ac9fbb