Struct rest::ctrl_io::InputKeywords

pub struct InputKeywords {

    pub print_level: usize,
    pub basis_path: String,
    pub basis_type: String,
    pub auxbas_path: String,
    pub auxbas_type: String,
    pub use_auxbas: bool,
    pub even_tempered_basis: bool,
    pub etb_start_atom_number: usize,
    pub etb_beta: f64,
    pub use_isdf: bool,
    pub isdf_k_mu: usize,
    pub eri_type: String,
    pub use_ri_symm: bool,
    pub xc: String,
    pub post_xc: Vec<String>,
    pub post_correlation: Vec<DFAFamily>,
    pub charge: f64,
    pub spin: f64,
    pub spin_channel: usize,
    pub spin_polarization: bool,
    pub frozen_core_postscf: i32,
    pub frequency_points: usize,
    pub lambda_points: usize,
    pub freq_grid_type: usize,
    pub freq_cut_off: f64,
    pub radial_precision: f64,
    pub min_num_angular_points: usize,
    pub max_num_angular_points: usize,
    pub grid_gen_level: usize,
    pub hardness: usize,
    pub pruning: String,
    pub rad_grid_method: String,
    pub external_grids: String,
    pub mixer: String,
    pub mix_param: f64,
    pub num_max_diis: usize,
    pub start_diis_cycle: usize,
    pub start_check_oscillation: usize,
    pub max_scf_cycle: usize,
    pub scf_acc_rho: f64,
    pub scf_acc_eev: f64,
    pub scf_acc_etot: f64,
    pub restart: bool,
    pub chkfile: String,
    pub chkfile_type: String,
    pub guessfile: String,
    pub guessfile_type: String,
    pub external_init_guess: bool,
    pub initial_guess: String,
    pub noiter: bool,
    pub check_stab: bool,
    pub use_dm_only: bool,
    pub fciqmc_dump: bool,
    pub outputs: Vec<String>,
    pub atom_sad: bool,
    pub num_threads: Option<usize>,

}

InputKeywords for a specific calculation

System dependent keywords

print_level: default (1). 0 dose not print anything. larger number with more output information

Basis set keywords

basis_path: String. The path where you can find the basis-set file in json format. If the basis-set file is missing, REST will try to download it from BasisSetExchange

basis_type: String. It can be spheric or cartesian

auxbas_path: String. The path where you can find the auxiliary basis-set file in json format. If the basis-set file is missing, REST will try to download it from BasisSetExchange

auxbas_type: String. It can be spheric or cartesian

even_tempered-basis: Bool. True: turn on ETB to generate the auxiliary basis set

etb_start_atom_number: Usize. Use ETB, for the element with atomic index larger than this value

etb_beta: f64. Relevant to the ETB basis set size. Smaller value indicates larger ETB basis set. NOTE: etb_beta should be larger than 1.0

Fields
print_level: usizebasis_path: Stringbasis_type: Stringauxbas_path: Stringauxbas_type: Stringuse_auxbas: booleven_tempered_basis: booletb_start_atom_number: usizeetb_beta: f64use_isdf: boolisdf_k_mu: usizeeri_type: Stringuse_ri_symm: boolxc: Stringpost_xc: Vec<String>post_correlation: Vec<DFAFamily>charge: f64spin: f64spin_channel: usizespin_polarization: boolfrozen_core_postscf: i32frequency_points: usizelambda_points: usizefreq_grid_type: usizefreq_cut_off: f64radial_precision: f64min_num_angular_points: usizemax_num_angular_points: usizegrid_gen_level: usizehardness: usizepruning: Stringrad_grid_method: Stringexternal_grids: Stringmixer: Stringmix_param: f64num_max_diis: usizestart_diis_cycle: usizestart_check_oscillation: usizemax_scf_cycle: usizescf_acc_rho: f64scf_acc_eev: f64scf_acc_etot: f64restart: boolchkfile: Stringchkfile_type: Stringguessfile: Stringguessfile_type: Stringexternal_init_guess: boolinitial_guess: Stringnoiter: boolcheck_stab: booluse_dm_only: boolfciqmc_dump: booloutputs: Vec<String>atom_sad: boolnum_threads: Option<usize>
Implementations

impl InputKeywords

pub fn new(
xc: String,
charge: f64,
spin: f64,
basis_path: String,
auxbas_path: String,
print_level: usize,
num_threads: usize,
basis_type: Option<String>
) -> PyResult<InputKeywords>

pub fn py_set_print_level(&mut self, print_level: usize)

pub fn py_set_xc(&mut self, xc: String)

pub fn py_set_basis_path(&mut self, basis_path: String)

pub fn py_set_eri_type(&mut self, eri_type: String)

pub fn py_set_basis_type(&mut self, basis_type: String)

pub fn py_set_auxbasis_path(&mut self, auxbasis_path: String)

pub fn py_set_auxbasis_type(&mut self, auxbasis_type: String)

pub fn py_set_charge_spin(&mut self, charge_and_spin: [f64; 2])

pub fn py_set_initial_guess(&mut self, init_guess: String)

pub fn py_set_spin_polarization(&mut self, flag: bool)

pub fn py_set_num_threads(&mut self, num_threads: usize)

pub fn py_set_mixer(&mut self, mixer: String)

pub fn py_set_mix_param(&mut self, mix_param: f64)

pub fn py_set_etb(&mut self, etb_flag: bool)

impl InputKeywords

pub fn init_ctrl() -> InputKeywords

pub fn parse_ctl_from_json(
tmp_keys: &Value
) -> Result<(InputKeywords, GeomCell)>

pub fn parse_ctl(filename: String) -> Result<(InputKeywords, GeomCell)>