Struct test_data_generation::data::field::Field

pub struct Field {
    pub name: String,
    pub value: Vec<u8>,
    // some fields omitted
}

Represents a Fact for a character in a sample data entity that has been analyzed

Fields
name: String

the textual name of the field

value: Vec<u8>

the value of the field

Methods

impl Field

fn new(name: String, field_type: FieldType) -> Field

Constructs a new Field

Example

extern crate test_data_generation;

use test_data_generation::data::field::{Field, FieldType};
    
fn main() {
    let mut fld = Field::new("field_1".to_string(), FieldType::Text);
        
    fld.set("John".to_string());
    println!("The value is {}",fld.get().downcast_ref::<String>().unwrap());
}

fn set<T: ConvertToBytes>(&mut self, val: T)

This function sets the value of the Field.

Example

extern crate test_data_generation;

use test_data_generation::data::field::{Field, FieldType};
    
fn main() {
    let mut fld = Field::new("first name".to_string(), FieldType::Text);
    fld.set("John".to_string());
}

fn get(&self) -> Box<Any>

This function returns the value for the Field Since a Box is return from this function, the actual value can be downcast to the proper type example: _.get().downcast_ref::().unwrap()_

Example

extern crate test_data_generation;

use test_data_generation::data::field::{Field, FieldType};
    
fn main() {
    let mut fld = Field::new("field_1".to_string(), FieldType::Text);
        
    fld.set("John".to_string());
    println!("The value is {}",fld.get().downcast_ref::<String>().unwrap());
}