Trait rmc_core::devices::motor_controllers::MotorController

pub trait MotorController: Send {
    fn set_speed(&mut self, new_speed: f32) -> Result<(), MotorFailure>;

    fn stop(&mut self) -> Result<(), MotorFailure>;

    fn invert(&mut self) -> Result<(), MotorFailure>;

    fn is_inverted(&self) -> Result<bool, MotorFailure>;
}

Required methods

fn set_speed(&mut self, new_speed: f32) -> Result<(), MotorFailure>

Sets the current speed of the motor controller. The speed should be a floating point number between -1 and 1. A negative speed indicates that the direction is reversed.

fn stop(&mut self) -> Result<(), MotorFailure>

Sets the current speed of the motor controller to zero.

fn invert(&mut self) -> Result<(), MotorFailure>

Inverts the directionality of the motor controller.

fn is_inverted(&self) -> Result<bool, MotorFailure>

Returns true if the motor controller is inverted and false otherwise.

Loading content...