Module rest::dft::gen_grids

Gen_grids is a library that produces numerical integration grid for molecules based on atom coordinates, atom types, and basis set information.
This work is based on Numgrid and PySCF.

Re-exports

pub use self::atom::atom_grid;

pub use self::atom::atom_grid_bse;

pub use self::lebedev::angular_grid;

pub use self::radial::radial_grid_kk;

pub use self::radial::radial_grid_lmg;

pub use self::radial::radial_grid_lmg_bse;

Modules

atom 🔒

Generate atomic grids based on given arguments.

becke_partitioning 🔒

Becke partitioning according to A. D. Becke. The Journal of Chemical Physics 88, 2547-2553 (1988). Reference can be found here.

bragg 🔒

Get bragg radius of the given atom.

bse 🔒

Download basis set information from www.basissetexchange.org to determine alpha_max and alpha_min.

comparison 🔒

lebedev 🔒

Generate Lebedev angular grids.

parameters 🔒

A set of constants used by gen_grids library.

prune 🔒

This mod provides methods to prune the number of angular grid.
Currently supported pruning methods:

python 🔒

radial 🔒

This mod is used to generate radial grids for DFT calculation.
Currently supported methods:

tables 🔒

Table for Lebedev angular grids.