Module rest::dft::gen_grids::atom

Generate atomic grids based on given arguments.

Functions

atom_grid

Generate atomic grids based on given arguments.

atom_grid_bse

Generate atomic grids based on given arguments.
Download basis set information from www.basissetexchange.org to determine alpha_max and alpha_min.

default_angular_num

Determine the number of angular grids according to level.

default_radial_num

Determine the number of radial grids according to level.

get_closest_n_ang 🔒

Determine the closest Lebedev grid number to the given number.