Module rest::dft::gen_grids::radial

This mod is used to generate radial grids for DFT calculation.
Currently supported methods:

Krack-Koster radial grid1

Lindh-Malmqvist-Gagliardi radial grid2

Treutler-Ahlrichs radial grid3

M. Krack, A. M. Köster. The Journal of Chemical Physics 108, 3226-3234 (1998). ↩

R. Lindh, P.-Å. Malmqvist, L. Gagliardi. Theoretical Chemistry Accounts 106, 178-187 (2001). ↩

O. Treutler, R. Ahlrichs. The Journal of Chemical Physics 102, 346-354 (1995). ↩

Functions

get_h 🔒

Lindh, Malmqvist, and Gagliardi radial grid according to R. Lindh, P.-Å. Malmqvist, L. Gagliardi. Theoretical Chemistry Accounts 106, 178-187 (2001), eqs.17-18.

get_r_inner 🔒

Lindh, Malmqvist, and Gagliardi radial grid according to R. Lindh, P.-Å. Malmqvist, L. Gagliardi. Theoretical Chemistry Accounts 106, 178-187 (2001), eq.25.

get_r_outer 🔒

Lindh, Malmqvist, and Gagliardi radial grid according to R. Lindh, P.-Å. Malmqvist, L. Gagliardi. Theoretical Chemistry Accounts 106, 178-187 (2001), eq.19.

kk_r_w 🔒

Generate Krack-Koster radial grid $r_i$ and weight $w_i$
M. Krack, A. M. Köster. The Journal of Chemical Physics 108, 3226-3234 (1998), eqs. 9-10.

radial_grid_becke

radial_grid_delley

radial_grid_gc2nd

Same as radial_grid_kk.

radial_grid_kk

Krack-Koster radial grid according to M. Krack, A. M. Köster. The Journal of Chemical Physics 108, 3226-3234 (1998), eqs. 9-13.
Reference can be found here.

radial_grid_lmg

Lindh, Malmqvist, and Gagliardi radial grid according to R. Lindh, P.-Å. Malmqvist, L. Gagliardi. Theoretical Chemistry Accounts 106, 178-187 (2001).
Reference can be found here.

radial_grid_lmg_bse

Download basis set information from www.basissetexchange.org to determine alpha_max and alpha_min, returning radial grids coordinates and weights.

radial_grid_mura_knowles

radial_grid_murray

radial_grid_treutler

Treutler-Ahlrichs radial grid according to O. Treutler, R. Ahlrichs. The Journal of Chemical Physics 102, 346-354 (1995).
Quadrature T2 with the mapping M4.
Reference can be found here.