Module rest::dft::libxc::ffi_xc

Structs

func_params_type

func_reference_type

max_align_t

xc_dimensions

xc_func_info_type

xc_func_type

Constants

XC_CORRELATION

XC_EXCHANGE

XC_EXCHANGE_CORRELATION

XC_EXT_PARAMS_DEFAULT

XC_FAMILY_GGA

XC_FAMILY_HYB_GGA

XC_FAMILY_HYB_LDA

XC_FAMILY_HYB_MGGA

XC_FAMILY_LCA

XC_FAMILY_LDA

XC_FAMILY_MGGA

XC_FAMILY_OEP

XC_FAMILY_UNKNOWN

XC_FLAGS_1D

XC_FLAGS_2D

XC_FLAGS_3D

XC_FLAGS_DEVELOPMENT

XC_FLAGS_HAVE_ALL

XC_FLAGS_HAVE_EXC

XC_FLAGS_HAVE_FXC

XC_FLAGS_HAVE_KXC

XC_FLAGS_HAVE_LXC

XC_FLAGS_HAVE_VXC

XC_FLAGS_HYB_CAM

XC_FLAGS_HYB_CAMY

XC_FLAGS_HYB_LC

XC_FLAGS_HYB_LCY

XC_FLAGS_NEEDS_LAPLACIAN

XC_FLAGS_STABLE

XC_FLAGS_VV10

XC_GGA_C_ACGGA

XC_GGA_C_ACGGAP

XC_GGA_C_AM05

XC_GGA_C_APBE

XC_GGA_C_BCGP

XC_GGA_C_BGCP

XC_GGA_C_BMK

XC_GGA_C_CCDF

XC_GGA_C_CHACHIYO

XC_GGA_C_CS1

XC_GGA_C_FT97

XC_GGA_C_GAM

XC_GGA_C_GAPC

XC_GGA_C_GAPLOC

XC_GGA_C_HCTH_A

XC_GGA_C_HYB_TAU_HCTH

XC_GGA_C_LM

XC_GGA_C_LYP

XC_GGA_C_LYPR

XC_GGA_C_MGGAC

XC_GGA_C_N12

XC_GGA_C_N12_SX

XC_GGA_C_OPTC

XC_GGA_C_OP_B88

XC_GGA_C_OP_G96

XC_GGA_C_OP_PBE

XC_GGA_C_OP_PW91

XC_GGA_C_OP_XALPHA

XC_GGA_C_P86

XC_GGA_C_P86VWN

XC_GGA_C_P86VWN_FT

XC_GGA_C_P86_FT

XC_GGA_C_PBE

XC_GGA_C_PBEFE

XC_GGA_C_PBEINT

XC_GGA_C_PBELOC

XC_GGA_C_PBE_JRGX

XC_GGA_C_PBE_MOL

XC_GGA_C_PBE_SOL

XC_GGA_C_PBE_VWN

XC_GGA_C_PW91

XC_GGA_C_Q2D

XC_GGA_C_REGTPSS

XC_GGA_C_REVTCA

XC_GGA_C_RGE2

XC_GGA_C_SCAN_E0

XC_GGA_C_SG4

XC_GGA_C_SOGGA11

XC_GGA_C_SOGGA11_X

XC_GGA_C_SPBE

XC_GGA_C_TAU_HCTH

XC_GGA_C_TCA

XC_GGA_C_TM_LYP

XC_GGA_C_TM_PBE

XC_GGA_C_VPBE

XC_GGA_C_W94

XC_GGA_C_WI

XC_GGA_C_WI0

XC_GGA_C_WL

XC_GGA_C_XPBE

XC_GGA_C_ZPBEINT

XC_GGA_C_ZPBESOL

XC_GGA_C_ZVPBEINT

XC_GGA_C_ZVPBELOC

XC_GGA_C_ZVPBESOL

XC_GGA_K_ABSP1

XC_GGA_K_ABSP2

XC_GGA_K_ABSP3

XC_GGA_K_ABSP4

XC_GGA_K_ABSR1

XC_GGA_K_ABSR2

XC_GGA_K_APBE

XC_GGA_K_APBEINT

XC_GGA_K_BALTIN

XC_GGA_K_DK

XC_GGA_K_ERNZERHOF

XC_GGA_K_EXP4

XC_GGA_K_FR_B88

XC_GGA_K_FR_PW86

XC_GGA_K_GDS08

XC_GGA_K_GE2

XC_GGA_K_GHDS10

XC_GGA_K_GHDS10R

XC_GGA_K_GOLDEN

XC_GGA_K_GP85

XC_GGA_K_GR

XC_GGA_K_LC94

XC_GGA_K_LGAP

XC_GGA_K_LGAP_GE

XC_GGA_K_LIEB

XC_GGA_K_LKT

XC_GGA_K_LLP

XC_GGA_K_LUDENA

XC_GGA_K_MEYER

XC_GGA_K_OL1

XC_GGA_K_OL2

XC_GGA_K_PBE2

XC_GGA_K_PBE3

XC_GGA_K_PBE4

XC_GGA_K_PEARSON

XC_GGA_K_PERDEW

XC_GGA_K_PG1

XC_GGA_K_RATIONAL_P

XC_GGA_K_REVAPBE

XC_GGA_K_REVAPBEINT

XC_GGA_K_TFVW

XC_GGA_K_TFVW_OPT

XC_GGA_K_THAKKAR

XC_GGA_K_TKVLN

XC_GGA_K_TW1

XC_GGA_K_TW2

XC_GGA_K_TW3

XC_GGA_K_TW4

XC_GGA_K_VJKS

XC_GGA_K_VSK

XC_GGA_K_VT84F

XC_GGA_K_VW

XC_GGA_K_YT65

XC_GGA_XC_B97

XC_GGA_XC_B97_1

XC_GGA_XC_B97_2

XC_GGA_XC_B97_3

XC_GGA_XC_B97_D

XC_GGA_XC_B97_GGA1

XC_GGA_XC_B97_K

XC_GGA_XC_BEEFVDW

XC_GGA_XC_EDF1

XC_GGA_XC_HCTH_93

XC_GGA_XC_HCTH_120

XC_GGA_XC_HCTH_147

XC_GGA_XC_HCTH_407

XC_GGA_XC_HCTH_407P

XC_GGA_XC_HCTH_P14

XC_GGA_XC_HCTH_P76

XC_GGA_XC_HLE16

XC_GGA_XC_KT1

XC_GGA_XC_KT2

XC_GGA_XC_KT3

XC_GGA_XC_LB

XC_GGA_XC_MOHLYP

XC_GGA_XC_MOHLYP2

XC_GGA_XC_MPWLYP1W

XC_GGA_XC_NCAP

XC_GGA_XC_OBLYP_D

XC_GGA_XC_OPBE_D

XC_GGA_XC_OPWLYP_D

XC_GGA_XC_PBE1W

XC_GGA_XC_PBELYP1W

XC_GGA_XC_SB98_1a

XC_GGA_XC_SB98_1b

XC_GGA_XC_SB98_1c

XC_GGA_XC_SB98_2a

XC_GGA_XC_SB98_2b

XC_GGA_XC_SB98_2c

XC_GGA_XC_TH1

XC_GGA_XC_TH2

XC_GGA_XC_TH3

XC_GGA_XC_TH4

XC_GGA_XC_TH_FC

XC_GGA_XC_TH_FCFO

XC_GGA_XC_TH_FCO

XC_GGA_XC_TH_FL

XC_GGA_XC_VV10

XC_GGA_XC_WB97

XC_GGA_XC_WB97X

XC_GGA_XC_WB97X_D

XC_GGA_XC_WB97X_V

XC_GGA_XC_XLYP

XC_GGA_X_2D_B86

XC_GGA_X_2D_B86_MGC

XC_GGA_X_2D_B88

XC_GGA_X_2D_PBE

XC_GGA_X_AIRY

XC_GGA_X_AK13

XC_GGA_X_AM05

XC_GGA_X_APBE

XC_GGA_X_B86

XC_GGA_X_B86_MGC

XC_GGA_X_B86_R

XC_GGA_X_B88

XC_GGA_X_B88M

XC_GGA_X_B88_6311G

XC_GGA_X_BAYESIAN

XC_GGA_X_BCGP

XC_GGA_X_BEEFVDW

XC_GGA_X_BGCP

XC_GGA_X_BPCCAC

XC_GGA_X_C09X

XC_GGA_X_CAP

XC_GGA_X_CHACHIYO

XC_GGA_X_DK87_R1

XC_GGA_X_DK87_R2

XC_GGA_X_EB88

XC_GGA_X_ECMV92

XC_GGA_X_EV93

XC_GGA_X_FD_LB94

XC_GGA_X_FD_REVLB94

XC_GGA_X_FT97_A

XC_GGA_X_FT97_B

XC_GGA_X_G96

XC_GGA_X_GAM

XC_GGA_X_GG99

XC_GGA_X_HCTH_A

XC_GGA_X_HERMAN

XC_GGA_X_HJS_B88

XC_GGA_X_HJS_B88_V2

XC_GGA_X_HJS_B97X

XC_GGA_X_HJS_PBE

XC_GGA_X_HJS_PBE_SOL

XC_GGA_X_HTBS

XC_GGA_X_ITYH

XC_GGA_X_ITYH_OPTX

XC_GGA_X_ITYH_PBE

XC_GGA_X_KGG99

XC_GGA_X_KT1

XC_GGA_X_LAG

XC_GGA_X_LAMBDA_CH_N

XC_GGA_X_LAMBDA_LO_N

XC_GGA_X_LAMBDA_OC2_N

XC_GGA_X_LB

XC_GGA_X_LBM

XC_GGA_X_LG93

XC_GGA_X_LSPBE

XC_GGA_X_LSRPBE

XC_GGA_X_LV_RPW86

XC_GGA_X_MB88

XC_GGA_X_MPBE

XC_GGA_X_MPW91

XC_GGA_X_N12

XC_GGA_X_NCAP

XC_GGA_X_OL2

XC_GGA_X_OPTB86B_VDW

XC_GGA_X_OPTB88_VDW

XC_GGA_X_OPTPBE_VDW

XC_GGA_X_OPTX

XC_GGA_X_PBE

XC_GGA_X_PBEA

XC_GGA_X_PBEFE

XC_GGA_X_PBEINT

XC_GGA_X_PBEK1_VDW

XC_GGA_X_PBEPOW

XC_GGA_X_PBETRANS

XC_GGA_X_PBE_JSJR

XC_GGA_X_PBE_MOL

XC_GGA_X_PBE_R

XC_GGA_X_PBE_SOL

XC_GGA_X_PBE_TCA

XC_GGA_X_PBEpow

XC_GGA_X_PW86

XC_GGA_X_PW91

XC_GGA_X_PW91_MOD

XC_GGA_X_Q1D

XC_GGA_X_Q2D

XC_GGA_X_REVSSB_D

XC_GGA_X_RGE2

XC_GGA_X_RPBE

XC_GGA_X_RPW86

XC_GGA_X_S12G

XC_GGA_X_SFAT

XC_GGA_X_SFAT_PBE

XC_GGA_X_SG4

XC_GGA_X_SOGGA

XC_GGA_X_SOGGA11

XC_GGA_X_SSB

XC_GGA_X_SSB_D

XC_GGA_X_SSB_SW

XC_GGA_X_VMT84_GE

XC_GGA_X_VMT84_PBE

XC_GGA_X_VMT_GE

XC_GGA_X_VMT_PBE

XC_GGA_X_WC

XC_GGA_X_WPBEH

XC_GGA_X_XPBE

XC_HYB_GGA_XC_APBE0

XC_HYB_GGA_XC_APF

XC_HYB_GGA_XC_B1LYP

XC_HYB_GGA_XC_B1PW91

XC_HYB_GGA_XC_B1WC

XC_HYB_GGA_XC_B3LYP

XC_HYB_GGA_XC_B3LYP3

XC_HYB_GGA_XC_B3LYP5

XC_HYB_GGA_XC_B3LYPS

XC_HYB_GGA_XC_B3LYP_MCM1

XC_HYB_GGA_XC_B3LYP_MCM2

XC_HYB_GGA_XC_B3LYPs

XC_HYB_GGA_XC_B3P86

XC_HYB_GGA_XC_B3PW91

XC_HYB_GGA_XC_B97

XC_HYB_GGA_XC_B97_1

XC_HYB_GGA_XC_B97_1P

XC_HYB_GGA_XC_B97_1p

XC_HYB_GGA_XC_B97_2

XC_HYB_GGA_XC_B97_3

XC_HYB_GGA_XC_B97_K

XC_HYB_GGA_XC_B5050LYP

XC_HYB_GGA_XC_BHANDH

XC_HYB_GGA_XC_BHANDHLYP

XC_HYB_GGA_XC_BLYP35

XC_HYB_GGA_XC_CAMH_B3LYP

XC_HYB_GGA_XC_CAMY_B3LYP

XC_HYB_GGA_XC_CAMY_BLYP

XC_HYB_GGA_XC_CAMY_PBEH

XC_HYB_GGA_XC_CAM_B3LYP

XC_HYB_GGA_XC_CAM_O3LYP

XC_HYB_GGA_XC_CAM_PBEH

XC_HYB_GGA_XC_CAM_QTP_00

XC_HYB_GGA_XC_CAM_QTP_01

XC_HYB_GGA_XC_CAM_QTP_02

XC_HYB_GGA_XC_CAP0

XC_HYB_GGA_XC_CASE21

XC_HYB_GGA_XC_EDF2

XC_HYB_GGA_XC_HAPBE

XC_HYB_GGA_XC_HFLYP

XC_HYB_GGA_XC_HJS_B88

XC_HYB_GGA_XC_HJS_B97X

XC_HYB_GGA_XC_HJS_PBE

XC_HYB_GGA_XC_HJS_PBE_SOL

XC_HYB_GGA_XC_HPBEINT

XC_HYB_GGA_XC_HSE03

XC_HYB_GGA_XC_HSE06

XC_HYB_GGA_XC_HSE12

XC_HYB_GGA_XC_HSE12S

XC_HYB_GGA_XC_HSE_SOL

XC_HYB_GGA_XC_KMLYP

XC_HYB_GGA_XC_LB07

XC_HYB_GGA_XC_LCY_BLYP

XC_HYB_GGA_XC_LCY_PBE

XC_HYB_GGA_XC_LC_BLYP

XC_HYB_GGA_XC_LC_BLYPR

XC_HYB_GGA_XC_LC_BLYP_EA

XC_HYB_GGA_XC_LC_BOP

XC_HYB_GGA_XC_LC_PBEOP

XC_HYB_GGA_XC_LC_QTP

XC_HYB_GGA_XC_LC_VV10

XC_HYB_GGA_XC_LC_WPBE

XC_HYB_GGA_XC_LC_WPBE08_WHS

XC_HYB_GGA_XC_LC_WPBEH_WHS

XC_HYB_GGA_XC_LC_WPBESOL_WHS

XC_HYB_GGA_XC_LC_WPBE_WHS

XC_HYB_GGA_XC_LRC_WPBE

XC_HYB_GGA_XC_LRC_WPBEH

XC_HYB_GGA_XC_MB3LYP_RC04

XC_HYB_GGA_XC_MCAM_B3LYP

XC_HYB_GGA_XC_MPW1K

XC_HYB_GGA_XC_MPW1LYP

XC_HYB_GGA_XC_MPW1PBE

XC_HYB_GGA_XC_MPW1PW

XC_HYB_GGA_XC_MPW3LYP

XC_HYB_GGA_XC_MPW3PW

XC_HYB_GGA_XC_MPWLYP1M

XC_HYB_GGA_XC_O3LYP

XC_HYB_GGA_XC_PBE0_13

XC_HYB_GGA_XC_PBE38

XC_HYB_GGA_XC_PBE50

XC_HYB_GGA_XC_PBEB0

XC_HYB_GGA_XC_PBEH

XC_HYB_GGA_XC_PBE_2X

XC_HYB_GGA_XC_PBE_MOL0

XC_HYB_GGA_XC_PBE_MOLB0

XC_HYB_GGA_XC_PBE_SOL0

XC_HYB_GGA_XC_QTP17

XC_HYB_GGA_XC_RCAM_B3LYP

XC_HYB_GGA_XC_REVB3LYP

XC_HYB_GGA_XC_SB98_1A

XC_HYB_GGA_XC_SB98_1B

XC_HYB_GGA_XC_SB98_1C

XC_HYB_GGA_XC_SB98_1a

XC_HYB_GGA_XC_SB98_1b

XC_HYB_GGA_XC_SB98_1c

XC_HYB_GGA_XC_SB98_2A

XC_HYB_GGA_XC_SB98_2B

XC_HYB_GGA_XC_SB98_2C

XC_HYB_GGA_XC_SB98_2a

XC_HYB_GGA_XC_SB98_2b

XC_HYB_GGA_XC_SB98_2c

XC_HYB_GGA_XC_TUNED_CAM_B3LYP

XC_HYB_GGA_XC_WB97

XC_HYB_GGA_XC_WB97X

XC_HYB_GGA_XC_WB97X_D

XC_HYB_GGA_XC_WB97X_D3

XC_HYB_GGA_XC_WB97X_V

XC_HYB_GGA_XC_WC04

XC_HYB_GGA_XC_WHPBE0

XC_HYB_GGA_XC_WP04

XC_HYB_GGA_XC_X3LYP

XC_HYB_GGA_XC_mPW1K

XC_HYB_GGA_XC_mPW1PW

XC_HYB_GGA_X_CAM_S12G

XC_HYB_GGA_X_CAM_S12H

XC_HYB_GGA_X_N12_SX

XC_HYB_GGA_X_S12H

XC_HYB_GGA_X_SOGGA11_X

XC_HYB_LDA_XC_BN05

XC_HYB_LDA_XC_CAM_LDA0

XC_HYB_LDA_XC_LDA0

XC_HYB_MGGA_XC_B0KCIS

XC_HYB_MGGA_XC_B86B95

XC_HYB_MGGA_XC_B88B95

XC_HYB_MGGA_XC_B94_HYB

XC_HYB_MGGA_XC_B98

XC_HYB_MGGA_XC_BB1K

XC_HYB_MGGA_XC_EDMGGAH

XC_HYB_MGGA_XC_LC_TMLYP

XC_HYB_MGGA_XC_M08_HX

XC_HYB_MGGA_XC_M08_SO

XC_HYB_MGGA_XC_M11

XC_HYB_MGGA_XC_MPW1B95

XC_HYB_MGGA_XC_MPW1KCIS

XC_HYB_MGGA_XC_MPWB1K

XC_HYB_MGGA_XC_MPWKCIS1K

XC_HYB_MGGA_XC_PBE1KCIS

XC_HYB_MGGA_XC_PW6B95

XC_HYB_MGGA_XC_PW86B95

XC_HYB_MGGA_XC_PWB6K

XC_HYB_MGGA_XC_REVTPSSH

XC_HYB_MGGA_XC_TPSS0

XC_HYB_MGGA_XC_TPSS1KCIS

XC_HYB_MGGA_XC_TPSSH

XC_HYB_MGGA_XC_WB97M_V

XC_HYB_MGGA_XC_X1B95

XC_HYB_MGGA_XC_XB1K

XC_HYB_MGGA_X_BMK

XC_HYB_MGGA_X_DLDF

XC_HYB_MGGA_X_JS18

XC_HYB_MGGA_X_M05

XC_HYB_MGGA_X_M05_2X

XC_HYB_MGGA_X_M06

XC_HYB_MGGA_X_M06_2X

XC_HYB_MGGA_X_M06_HF

XC_HYB_MGGA_X_M06_SX

XC_HYB_MGGA_X_M08_HX

XC_HYB_MGGA_X_M08_SO

XC_HYB_MGGA_X_M11

XC_HYB_MGGA_X_MN12_SX

XC_HYB_MGGA_X_MN15

XC_HYB_MGGA_X_MS2H

XC_HYB_MGGA_X_MVSH

XC_HYB_MGGA_X_PJS18

XC_HYB_MGGA_X_REVM06

XC_HYB_MGGA_X_REVM11

XC_HYB_MGGA_X_REVSCAN0

XC_HYB_MGGA_X_SCAN0

XC_HYB_MGGA_X_TAU_HCTH

XC_KINETIC

XC_LDA_C_1D_CSC

XC_LDA_C_1D_LOOS

XC_LDA_C_2D_AMGB

XC_LDA_C_2D_PRM

XC_LDA_C_BR78

XC_LDA_C_CHACHIYO

XC_LDA_C_CHACHIYO_MOD

XC_LDA_C_GK72

XC_LDA_C_GL

XC_LDA_C_GOMBAS

XC_LDA_C_HL

XC_LDA_C_KARASIEV

XC_LDA_C_KARASIEV_MOD

XC_LDA_C_LP96

XC_LDA_C_LP_A

XC_LDA_C_LP_B

XC_LDA_C_MCWEENY

XC_LDA_C_ML1

XC_LDA_C_ML2

XC_LDA_C_OB_PW

XC_LDA_C_OB_PZ

XC_LDA_C_OW

XC_LDA_C_OW_LYP

XC_LDA_C_PK09

XC_LDA_C_PMGB06

XC_LDA_C_PW

XC_LDA_C_PW_MOD

XC_LDA_C_PW_RPA

XC_LDA_C_PZ

XC_LDA_C_PZ_MOD

XC_LDA_C_RC04

XC_LDA_C_RPA

XC_LDA_C_RPW92

XC_LDA_C_UPW92

XC_LDA_C_VBH

XC_LDA_C_VWN

XC_LDA_C_VWN_1

XC_LDA_C_VWN_2

XC_LDA_C_VWN_3

XC_LDA_C_VWN_4

XC_LDA_C_VWN_RPA

XC_LDA_C_W20

XC_LDA_C_WIGNER

XC_LDA_C_XALPHA

XC_LDA_C_vBH

XC_LDA_K_LP

XC_LDA_K_LP96

XC_LDA_K_TF

XC_LDA_K_ZLP

XC_LDA_X

XC_LDA_XC_1D_EHWLRG_1

XC_LDA_XC_1D_EHWLRG_2

XC_LDA_XC_1D_EHWLRG_3

XC_LDA_XC_CORRKSDT

XC_LDA_XC_GDSMFB

XC_LDA_XC_KSDT

XC_LDA_XC_LP_A

XC_LDA_XC_LP_B

XC_LDA_XC_TETER93

XC_LDA_XC_TIH

XC_LDA_XC_ZLP

XC_LDA_X_1D

XC_LDA_X_1D_EXPONENTIAL

XC_LDA_X_1D_SOFT

XC_LDA_X_2D

XC_LDA_X_ERF

XC_LDA_X_RAE

XC_LDA_X_REL

XC_LDA_X_SLOC

XC_LDA_X_YUKAWA

XC_MAX_REFERENCES

XC_MGGA_C_B88

XC_MGGA_C_B94

XC_MGGA_C_BC95

XC_MGGA_C_CC06

XC_MGGA_C_CS

XC_MGGA_C_DLDF

XC_MGGA_C_HLTAPW

XC_MGGA_C_KCIS

XC_MGGA_C_KCISK

XC_MGGA_C_LP90

XC_MGGA_C_M05

XC_MGGA_C_M05_2X

XC_MGGA_C_M06

XC_MGGA_C_M06_2X

XC_MGGA_C_M06_HF

XC_MGGA_C_M06_L

XC_MGGA_C_M06_SX

XC_MGGA_C_M08_HX

XC_MGGA_C_M08_SO

XC_MGGA_C_M11

XC_MGGA_C_M11_L

XC_MGGA_C_MN12_L

XC_MGGA_C_MN12_SX

XC_MGGA_C_MN15

XC_MGGA_C_MN15_L

XC_MGGA_C_PKZB

XC_MGGA_C_R2SCAN

XC_MGGA_C_R2SCAN01

XC_MGGA_C_R2SCANL

XC_MGGA_C_REVM06

XC_MGGA_C_REVM06_L

XC_MGGA_C_REVM11

XC_MGGA_C_REVSCAN

XC_MGGA_C_REVSCAN_VV10

XC_MGGA_C_REVTM

XC_MGGA_C_REVTPSS

XC_MGGA_C_RMGGAC

XC_MGGA_C_RPPSCAN

XC_MGGA_C_RREGTM

XC_MGGA_C_RSCAN

XC_MGGA_C_SCAN

XC_MGGA_C_SCANL

XC_MGGA_C_SCANL_RVV10

XC_MGGA_C_SCANL_VV10

XC_MGGA_C_SCAN_RVV10

XC_MGGA_C_SCAN_VV10

XC_MGGA_C_TM

XC_MGGA_C_TPSS

XC_MGGA_C_TPSSLOC

XC_MGGA_C_VSXC

XC_MGGA_K_CSK1

XC_MGGA_K_CSK4

XC_MGGA_K_CSK_LOC1

XC_MGGA_K_CSK_LOC4

XC_MGGA_K_GEA2

XC_MGGA_K_GEA4

XC_MGGA_K_L04

XC_MGGA_K_L06

XC_MGGA_K_PC07

XC_MGGA_K_PC07_OPT

XC_MGGA_K_PGSL025

XC_MGGA_K_RDA

XC_MGGA_XC_B97M_V

XC_MGGA_XC_CC06

XC_MGGA_XC_HLE17

XC_MGGA_XC_LP90

XC_MGGA_XC_OTPSS_D

XC_MGGA_XC_TPSSLYP1W

XC_MGGA_XC_ZLP

XC_MGGA_X_2D_JS17

XC_MGGA_X_2D_PRHG07

XC_MGGA_X_2D_PRHG07_PRP10

XC_MGGA_X_B00

XC_MGGA_X_BJ06

XC_MGGA_X_BLOC

XC_MGGA_X_BR89

XC_MGGA_X_BR89_1

XC_MGGA_X_BR89_EXPLICIT

XC_MGGA_X_BR89_EXPLICIT_1

XC_MGGA_X_EDMGGA

XC_MGGA_X_FT98

XC_MGGA_X_GDME_0

XC_MGGA_X_GDME_KOS

XC_MGGA_X_GDME_NV

XC_MGGA_X_GDME_VT

XC_MGGA_X_GVT4

XC_MGGA_X_GX

XC_MGGA_X_HLTA

XC_MGGA_X_JK

XC_MGGA_X_LTA

XC_MGGA_X_M05

XC_MGGA_X_M05_2X

XC_MGGA_X_M06

XC_MGGA_X_M06_2X

XC_MGGA_X_M06_HF

XC_MGGA_X_M06_L

XC_MGGA_X_M08_HX

XC_MGGA_X_M08_SO

XC_MGGA_X_M11

XC_MGGA_X_M11_L

XC_MGGA_X_MBEEF

XC_MGGA_X_MBEEFVDW

XC_MGGA_X_MBR

XC_MGGA_X_MBRXC_BG

XC_MGGA_X_MBRXH_BG

XC_MGGA_X_MCML

XC_MGGA_X_MGGAC

XC_MGGA_X_MK00

XC_MGGA_X_MK00B

XC_MGGA_X_MN12_L

XC_MGGA_X_MN12_SX

XC_MGGA_X_MN15_L

XC_MGGA_X_MODTPSS

XC_MGGA_X_MS0

XC_MGGA_X_MS1

XC_MGGA_X_MS2

XC_MGGA_X_MS2B

XC_MGGA_X_MS2BS

XC_MGGA_X_MS2_REV

XC_MGGA_X_MTASK

XC_MGGA_X_MVS

XC_MGGA_X_MVSB

XC_MGGA_X_MVSBS

XC_MGGA_X_PBE_GX

XC_MGGA_X_PKZB

XC_MGGA_X_R2SCAN

XC_MGGA_X_R2SCAN01

XC_MGGA_X_R2SCANL

XC_MGGA_X_R4SCAN

XC_MGGA_X_REGTM

XC_MGGA_X_REGTPSS

XC_MGGA_X_REVM06_L

XC_MGGA_X_REVSCAN

XC_MGGA_X_REVSCANL

XC_MGGA_X_REVTM

XC_MGGA_X_REVTPSS

XC_MGGA_X_RLDA

XC_MGGA_X_RPP09

XC_MGGA_X_RPPSCAN

XC_MGGA_X_RSCAN

XC_MGGA_X_RTPSS

XC_MGGA_X_SA_TPSS

XC_MGGA_X_SCAN

XC_MGGA_X_SCANL

XC_MGGA_X_TASK

XC_MGGA_X_TAU_HCTH

XC_MGGA_X_TB09

XC_MGGA_X_TH

XC_MGGA_X_TLDA

XC_MGGA_X_TM

XC_MGGA_X_TPSS

XC_MGGA_X_VT84

XC_NON_RELATIVISTIC

XC_POLARIZED

XC_RELATIVISTIC

XC_TAU_EXPANSION

XC_TAU_EXPLICIT

XC_UNPOLARIZED

Functions

xc_aux_func_ids⚠

xc_aux_func_weights⚠

xc_available_functional_names⚠

Fills the list with the names of the available functionals, ordered by name. The list array should be [Nfuncs][maxlen+1].

xc_available_functional_numbers⚠

Returns the available functional number sorted by id

xc_available_functional_numbers_by_name⚠

Returns the available functional number sorted by the functionals’ names; this function is a helper for the Python frontend.

xc_family_from_id⚠

Get a functional’s family and the number within the family from the id number

xc_func_alloc⚠

Dynamically allocates a libxc functional; which will also need to be initialized.

xc_func_end⚠

Destructor for an initialized functional

xc_func_free⚠

Frees a dynamically allocated functional

xc_func_get_info⚠

Get information on a functional

xc_func_info_get_ext_params_default_value⚠

xc_func_info_get_ext_params_description⚠

xc_func_info_get_ext_params_name⚠

xc_func_info_get_family⚠

xc_func_info_get_flags⚠

xc_func_info_get_kind⚠

xc_func_info_get_n_ext_params⚠

xc_func_info_get_name⚠

xc_func_info_get_number⚠

xc_func_info_get_references⚠

xc_func_init⚠

Initializes a functional by id with nspin spin channels

xc_func_reference_get_bibtex⚠

xc_func_reference_get_doi⚠

xc_func_reference_get_ref⚠

xc_func_set_dens_threshold⚠

Sets the density threshold for a functional

xc_func_set_ext_params⚠

Sets all external parameters for a functional

xc_func_set_ext_params_name⚠

Sets an external parameter by name for a functional

xc_func_set_sigma_threshold⚠

Sets the reduced gradient threshold for a functional

xc_func_set_tau_threshold⚠

Sets the kinetic energy density threshold for a functional

xc_func_set_zeta_threshold⚠

Sets the spin polarization threshold for a functional

xc_functional_get_name⚠

Get a functional’s name from its id number

xc_functional_get_number⚠

Get a functional’s id number from its name

xc_gga⚠

Evaluate a GGA functional

xc_gga_ak13_get_asymptotic⚠

xc_gga_ak13_pars_get_asymptotic⚠

xc_gga_exc⚠

Evaluates the energy density for a GGA functional

xc_gga_exc_vxc⚠

Evaluates the energy density and its first derivative for a GGA functional

xc_gga_exc_vxc_fxc⚠

Evaluates the energy density and its first and second derivatives for a GGA functional

xc_gga_exc_vxc_fxc_kxc⚠

Evaluates the energy density and its first, second, and third derivatives for a GGA functional

xc_gga_fxc⚠

Evaluates the second derivative for a GGA functional

xc_gga_kxc⚠

Evaluates the third derivative for a GGA functional

xc_gga_lxc⚠

Evaluates the fourth derivative for a GGA functional

xc_gga_vxc⚠

Evaluates the first derivative of the energy density for a GGA functional

xc_gga_vxc_fxc⚠

Evaluates the first and second derivatives for a GGA functional

xc_gga_vxc_fxc_kxc⚠

Evaluates the first, second, and third derivatives for a GGA functional

xc_hyb_cam_coef⚠

xc_hyb_exx_coef⚠

Get the hybrid coefficient

xc_lda⚠

Evaluate an LDA functional

xc_lda_exc⚠

Evaluates the energy density for an LDA functional

xc_lda_exc_vxc⚠

Evaluates the energy density and its first derivative for an LDA functional

xc_lda_exc_vxc_fxc⚠

Evaluates the energy density and its first and second derivatives for an LDA functional

xc_lda_exc_vxc_fxc_kxc⚠

Evaluates the energy density and its first, second, and third derivatives for an LDA functional

xc_lda_fxc⚠

Evaluates the second derivative for an LDA functional

xc_lda_kxc⚠

Evaluates the third derivative for an LDA functional

xc_lda_lxc⚠

Evaluates the fourth derivative for an LDA functional

xc_lda_vxc⚠

Evaluates the first derivative of the energy density for an LDA functional

xc_lda_vxc_fxc⚠

Evaluates the first and second derivatives for an LDA functional

xc_lda_vxc_fxc_kxc⚠

Evaluates the first, second, and third derivatives for an LDA functional

xc_maximum_name_length⚠

The maximum name length of any functional

xc_mgga⚠

Evaluate a meta-GGA functional

xc_mgga_exc⚠

Evaluates the energy density for a meta-GGA functional

xc_mgga_exc_vxc⚠

Evaluates the energy density and its first derivative for a meta-GGA functional

xc_mgga_exc_vxc_fxc⚠

Evaluates the energy density and its first and second derivatives for a meta-GGA functional

xc_mgga_exc_vxc_fxc_kxc⚠

Evaluates the energy density and its first, second, and third derivatives for a meta-GGA functional

xc_mgga_fxc⚠

Evaluates the second derivative for a meta-GGA functional

xc_mgga_kxc⚠

Evaluates the third derivative for a meta-GGA functional

xc_mgga_lxc⚠

Evaluates the fourth derivative for a meta-GGA functional

xc_mgga_vxc⚠

Evaluates the first derivative of the energy density for a meta-GGA functional

xc_mgga_vxc_fxc⚠

Evaluates the first and second derivatives for a meta-GGA functional

xc_mgga_vxc_fxc_kxc⚠

Evaluates the first, second, and third derivatives for a meta-GGA functional

xc_nlc_coef⚠

xc_num_aux_funcs⚠

xc_number_of_functionals⚠

The number of functionals implemented in this version of libxc

xc_reference⚠

xc_reference_doi⚠

xc_version⚠

xc_version_string⚠

Type Definitions

size_t

wchar_t