layers
Module drawing

Drawing related code for Layer, Scene, and Text. The functions in this module are used to draw into a Skia canvas independent from the backend

Functions

draw_layer

Draw a layer into a skia::Canvas. Returns the damage rect in the layer’s coordinate space.

draw_scene

Draw the scene to the given skia::Canvas

node_tree_list

node_tree_list_visible

paint_node_tree

paint a node and his subtree in the provided canvas

print_scene

Print the node tree to the console

render_node_tree

Render a node and his subtree in the provided canvas taking care of the image caching if needed