Trait algo::ecdsa::EcdsaKeyGen

pub trait EcdsaKeyGen: ErrorType {
    type PrivateKey;
    type PublicKey;
    type Curve: EcdsaCurve;

    // Required method
    fn generate_key_pair(
        curve: &Self::Curve,
    ) -> Result<(Self::PrivateKey, Self::PublicKey), Self::Error>;
}

Trait for ECDSA key generation.

This trait defines the methods required for generating ECDSA key pairs.

Required Associated Types

type PrivateKey

type PublicKey

type Curve: EcdsaCurve

Required Methods

fn generate_key_pair( curve: &Self::Curve, ) -> Result<(Self::PrivateKey, Self::PublicKey), Self::Error>

Generates an ECDSA key pair.

Parameters

curve: The elliptic curve to use for key generation.

Returns

A result containing the generated private and public keys, or an error.

Object Safety

This trait is not object safe.