Module feedpress::endpoints

Functions

api_get_config

api_get_edition_list

api_get_logs

GET request will return the contents of the log file for display on the UI or access elsewhere.

api_get_version

GET request will return a version string, located in the env variable VERSION

api_press_edition

api_remove_edition

api_update_config

get_file_create

Returns a formatted string containing the file’s creation date or “01011900” if unable to determine.

get_file_name

Returns a formatted filename or the string “null” if unable to determine.

get_file_pages

Returns a simple numeric value of the pages in the PDF, based on what lopdf gives us from its metadata

get_file_size

Returns a formatted string containing a plain english representation of a file’s size. (eg: “125k”)

get_img_name

Returns the image file that corresponds to the input file. (strips the extension and path and then just returns it as .png)