layers
Module engine

Contains all the logic for running the animations, execute the scheduled changes to the models, and prepare the render tree The scene is drawn in 4 stages:

The layout step calculates the dimensions and position of a node and generates a transformation Matrix

The draw step generates a displaylist

The render step uses the displaylist to generate a texture of the node

The compose step generates the final image using the textures

The Engine is the main engine responsible for managing layers and rendering.

Usage:

use layers::prelude::*;

let engine = Engine::create(500.0, 500.0);
let layer = engine.new_layer();
engine.add_layer(&layer);

Modules

animation

Timing functions for animations:

Structs

AnimatedNodeChange

AnimationCallback

AnimationRef

AnimationState

A struct that contains the state of a given animation.

Engine

Public API for the Layers Engine

HandlerRef

NodeRef

An identifier for a node in the three storage

PointerCallback

SceneNode

Contains the layout of a layer and information required for drawing

Timestamp

TransactionCallback

TransactionRef

Enums

PointerEventType

TransactionEventType

Traits

Command

A trait for objects that can be exectuded by the engine.

SyncCommand