layers::engine
Struct SceneNode

pub struct SceneNode { /* private fields */ }

Contains the layout of a layer and information required for drawing

Implementations

impl SceneNode

pub fn new() -> Self

pub fn insert_flags(&mut self, flags: RenderableFlags)

pub fn remove_flags(&mut self, flags: RenderableFlags)

pub fn bounds(&self) -> Rect

pub fn bounds_with_children(&self) -> Rect

pub fn transformed_bounds(&self) -> Rect

pub fn transformed_bounds_with_effects(&self) -> Rect

pub fn transform(&self) -> Matrix

pub fn mark_for_deletion(&mut self)

pub fn is_deleted(&self) -> bool

pub fn set_debug_info(&mut self, debug_info: bool)

pub fn set_hidden(&mut self, hidden: bool)

pub fn hidden(&self) -> bool

pub fn set_image_cached(&mut self, value: bool)

pub fn is_image_cached(&self) -> bool

pub fn set_picture_cached(&mut self, value: bool)

pub fn is_picture_cached(&self) -> bool

pub fn set_content_opaque(&mut self, value: bool)

Hint that the custom draw content fills the entire bounds with opaque pixels. When set, the layer can act as an occluder for occlusion culling even when its background color is transparent.

pub fn is_content_opaque(&self) -> bool

pub fn render_layer(&self) -> &RenderLayer

pub fn render_layer_mut(&mut self) -> &mut RenderLayer

Mutable access to the render layer for testing and diagnostic tools.

pub fn set_needs_repaint(&mut self, need_repaint: bool)

pub fn set_needs_layout(&mut self, need_layout: bool)

pub fn needs_repaint(&self) -> bool

pub fn needs_layout(&self) -> bool

pub fn pointer_events(&self) -> bool

pub fn contains_point(&self, point: &Point) -> bool

pub fn frame_number(&self) -> usize

Returns the current frame number for this node, incremented each time the node is repainted.