Function rust_action_heroes::entity::make_camera

pub(crate) fn make_camera(
    world: &mut World, 
    level_handle: &Handle<GameLevel>
) -> Entity

Initializes the camera in the level.

It creates a camera with 1270x720 dimensions except we divide both of those numbers by 8 since our sprite are so small.

We also center the camera on based on the level dimenisons, othwerise the level would always have the bottom left corner on the center of the screen and the 3/4 of the screen would be empty.