Enum keychain_services::ErrorKind

pub enum ErrorKind {
    AuthFailed,
    BufferTooSmall,
    CreateChainFailed,
    DataTooLarge,
    DataNotAvailable,
    DataNotModifiable,
    DuplicateCallback,
    DuplicateItem,
    DuplicateKeychain,
    InDarkWake,
    InteractionNotAllowed,
    InteractionRequired,
    InvalidCallback,
    InvalidItemRef,
    InvalidKeychain,
    InvalidPrefsDomain,
    InvalidSearchRef,
    ItemNotFound,
    KeySizeNotAllowed,
    MissingEntitlement,
    NoCertificateModule,
    NoDefaultKeychain,
    NoPolicyModule,
    NoStorageModule,
    NoSuchAttr,
    NoSuchClass,
    NoSuchKeychain,
    NotAvailable,
    ReadOnly,
    ReadOnlyAttr,
    WrongSecVersion,
    Io {
        kind: ErrorKind,
    },
    CFError {
        code: i64,
        domain: String,
    },
    Errno {
        code: u8,
    },
    OSError {
        code: i64,
    },
}

Kinds of errors.

Variants
AuthFailed

Authentication and/or authorization failed.

Wrapper for the errSecAuthFailed status code. See: https://developer.apple.com/documentation/security/errsecauthfailed

BufferTooSmall

Buffer is too small.

Wrapper for the errSecBufferTooSmall status code. See: https://developer.apple.com/documentation/security/errsecbuffertoosmall

CreateChainFailed

Certificate chain creation attempt failed.

Wrapper for the errSecCreateChainFailed status code. See: https://developer.apple.com/documentation/security/errseccreatechainfailed

DataTooLarge

Data too large for the given data type.

Wrapper for the errSecDataTooLarge status code. See: https://developer.apple.com/documentation/security/errsecdatatoolarge

DataNotAvailable

Data is not available.

Wrapper for the errSecDataNotAvailable status code. See: https://developer.apple.com/documentation/security/errsecdatanotavailable

DataNotModifiable

Data cannot be modified.

Wrapper for the errSecDataNotModifiable status code. See: https://developer.apple.com/documentation/security/errsecdatanotmodifiable

DuplicateCallback

Callback with the same name already exists.

Wrapper for the errSecDuplicateCallback status code. See: https://developer.apple.com/documentation/security/errsecduplicatecallback

DuplicateItem

Item already exists.

Wrapper for the errSecDuplicateItem status code. See: https://developer.apple.com/documentation/security/errsecduplicateitem

DuplicateKeychain

Keychain with the same name already exists.

Wrapper for the errSecDuplicateKeychain status code. See: https://developer.apple.com/documentation/security/errsecduplicatekeychain

InDarkWake

System is in a dark wake state - user interface cannot be displayed.

Wrapper for the errSecInDarkWake status code. See: https://developer.apple.com/documentation/security/errsecindarkwake

InteractionNotAllowed

Security Server interactions not allowed in this context.

Wrapper for the errSecInteractionNotAllowed status code. See: https://developer.apple.com/documentation/security/errsecinteractionnotallowed

InteractionRequired

User interaction required.

Wrapper for the errSecInteractionRequired status code. See: https://developer.apple.com/documentation/security/errsecinteractionrequired

InvalidCallback

Callback is invalid.

Wrapper for the errSecInvalidCallback status code. See: https://developer.apple.com/documentation/security/errsecinvalidcallback

InvalidItemRef

Item reference is invalid.

Wrapper for the errSecInvalidItemRef status code. See: https://developer.apple.com/documentation/security/errsecinvaliditemref

InvalidKeychain

Keychain is invalid.

Wrapper for the errSecInvalidKeychain status code. See: https://developer.apple.com/documentation/security/errsecinvalidkeychain

InvalidPrefsDomain

Specified preference domain is not valid.

Wrapper for the errSecInvalidPrefsDomain status code. See: https://developer.apple.com/documentation/security/errsecinvalidprefsdomain

InvalidSearchRef

Search reference is invalid.

Wrapper for the errSecInvalidSearchRef status code. See: https://developer.apple.com/documentation/security/errsecinvalidsearchref

ItemNotFound

Item could not be found.

Wrapper for the errSecItemNotFound status code. See: https://developer.apple.com/documentation/security/errsecitemnotfound

KeySizeNotAllowed

Invalid key size.

Wrapper for the errSecKeySizeNotAllowed status code. See: https://developer.apple.com/documentation/security/errseckeysizenotallowed

MissingEntitlement

Required entitlement for accessing the keychain is missing. This error occurs when attempting to access certain keychain functionality from an application which is either unsigned or missing a required entitlement.

Wrapper for the errSecMissingEntitlement status code. See: https://developer.apple.com/documentation/security/errsecmissingentitlement

NoCertificateModule

Certificate module unavailable.

Wrapper for the errSecNoCertificateModule status code. See: https://developer.apple.com/documentation/security/errsecnocertificatemodule

NoDefaultKeychain

Default keychain does not exist.

Wrapper for the errSecNoDefaultKeychain status code. See: https://developer.apple.com/documentation/security/errsecnodefaultkeychain

NoPolicyModule

Policy module unavailable.

Wrapper for the errSecNoPolicyModule status code. See: https://developer.apple.com/documentation/security/errsecnopolicymodule

NoStorageModule

Storage module unavailable.

Wrapper for the errSecNoStorageModule status code. See: https://developer.apple.com/documentation/security/errsecnostoragemodule

NoSuchAttr

Specified attribute does not exist.

Wrapper for the errSecNoSuchAttr status code. See:; https://developer.apple.com/documentation/security/errsecnosuchattr

NoSuchClass

Specified keychain item class does not exist.

Wrapper for the errSecNoSuchClass status code. See: https://developer.apple.com/documentation/security/errsecnosuchclass

NoSuchKeychain

Specified keychain does not exist.

Wrapper for the errSecNoSuchKeychain status code. See: https://developer.apple.com/documentation/security/errsecnosuchkeychain

NotAvailable

Trust results not available.

Wrapper for the errSecNotAvailable status code. See: https://developer.apple.com/documentation/security/errsecnotavailable

ReadOnly

Can't perform given action on read-only item.

Wrapper for the errSecReadOnly status code. See: https://developer.apple.com/documentation/security/errsecreadonly

ReadOnlyAttr

Can't perform action on read-only attribute

Wrapper for the errSecReadOnlyAttr status code. See: https://developer.apple.com/documentation/security/errsecreadonlyattr

WrongSecVersion

Invalid version.

Wrapper for the errSecWrongSecVersion status code. See: https://developer.apple.com/documentation/security/errsecwrongversion

Io

Input/output errors.

Wrapper for errno codes we know/commonly encounter.

Fields of Io

kind: ErrorKind

std::io::ErrorKind value representing the I/O error

CFError

Errors returned from CoreFoundation.

Codes correspond to the return value of the CFErrorGetCode function.

For more information, see: https://developer.apple.com/documentation/corefoundation/1494656-cferrorgetcode?language=objc

Fields of CFError

code: i64

Code identifying this type of CFError.

See CFErrorGetCode() for more information: https://developer.apple.com/documentation/corefoundation/1494656-cferrorgetcode

domain: String

Domain associated with this error.

See CFErrorGetDomain() for more information: https://developer.apple.com/documentation/corefoundation/1494657-cferrorgetdomain

Errno

Unix errno values we receive from the underlying OS

Fields of Errno

code: u8

Raw errno value

OSError

OSStatus codes which we can't otherwise decode.

Fields of OSError

code: i64

OS error code