Enum grace_lib::expression::Stmt

pub enum Stmt {
    AssignmentStmt {
        name: Identifier,
        expression: Node<Expr>,
    },
    LetStmt {
        name: Identifier,
        type_annotation: Option<Type>,
        expression: Node<Expr>,
    },
    FunctionDecStmt {
        name: Identifier,
        args: Vec<(Identifier, Type)>,
        kwargs: Vec<(Identifier, Type, Node<Expr>)>,
        block: Node<Block>,
        return_type: Type,
    },
    StructDec {
        name: Identifier,
        fields: Vec<(Identifier, Type)>,
    },
    IfStmt {
        condition: Node<Expr>,
        block: Node<Block>,
        else_block: Option<Node<Block>>,
    },
    WhileStmt {
        condition: Node<Expr>,
        block: Node<Block>,
    },
    ReturnStmt(Node<Expr>),
    YieldStmt(Node<Expr>),
    BreakStmt,
    PassStmt,
    ContinueStmt,
}

Variants

AssignmentStmt

Fields

name: Identifier

expression: Node<Expr>

LetStmt

Fields

name: Identifier

type_annotation: Option<Type>

expression: Node<Expr>

FunctionDecStmt

Fields

name: Identifier

args: Vec<(Identifier, Type)>

kwargs: Vec<(Identifier, Type, Node<Expr>)>

block: Node<Block>

return_type: Type

StructDec

Fields

name: Identifier

fields: Vec<(Identifier, Type)>

IfStmt

Fields

condition: Node<Expr>

block: Node<Block>

else_block: Option<Node<Block>>

WhileStmt

Fields

condition: Node<Expr>

block: Node<Block>

ReturnStmt(Node<Expr>)

YieldStmt(Node<Expr>)

BreakStmt

PassStmt

ContinueStmt

Implementations

impl Stmt

pub fn simple_if<T>(condition: T, block: Block) -> Stmtwhere Node<Expr>: From<T>,

Construct an if statement from a condition and a block.

pub fn assn<T, S>(name: T, expr: S) -> Stmtwhere Identifier: From<T>, Expr: From<S>,

Construct an assignment statement from an identifier and an expression.

pub fn no_args<T>(name: T, block: Block, ret: Type) -> Stmtwhere Identifier: From<T>,

Construct a no-argument function declaration from a name, a block, and a return type.

pub fn ret<T>(expr: T) -> Stmtwhere Expr: From<T>,

Construct a return statement from an expression.

impl Stmt

pub fn get_name(&self) -> Identifier