Module wasm_svg_graphics::figures

Container for the figures, this includes all definitions for shapes and styling This is the module containing all the logic for shapes and styling

Modules

circle
Module containing CircleProps (The properties used when creating a Shape::Circle)

path
Module containing PathProps (The properties used when creating a Shape::Path)

preset
A set of presets for Figure, e.g. lines, circles, ...

shape
Module containing the definition for Shape, ShapeStyle, AttributeField

sub_path
Module containing the definition of SubPath which is used for defining smaller parts of a whole Shape::Path

Structs

Figure
A combination of shapes into one object used as a svg-def