Module wirefish::filtering

Retrieve packets filtered by their attributes

Available filters

By Protocol

UNKNOWN

ETHERNET

IPV4

IPV6

ARP

ICMP

ICMPV6

TCP

UDP

TLS

DNS

HTTP

By Attributes

SOURCE MAC

DESTINATION MAC

SOURCE IP

DESTINATION IP

SOURCE PORT

DESTINATION PORT

By Type

MALFORMED

Re-exports

pub use __cmd__get_packets;

Modules

FilterNamesValues 🔒

Structs

PacketsCollection

List of all the collected packets and additional data structures to speed up the filtering process

Functions

apply_all_strong_filters 🔒

apply_layer_type_filter

Filter collected packets by protocol type

apply_specific_filter

Filter collected packets by packet attribute

bound_and_map 🔒

filter_by_dst_ip

Filter collected packets by destination IP address

filter_by_dst_mac

Filter collected packets by destination MAC address

filter_by_dst_port

Filter collected packets by destination Port

filter_by_src_ip

Filter collected packets by Soure IP address

filter_by_src_mac

Filter collected packets by source MAC address

filter_by_src_port

Filter collected packets by source Port

get_bounded_type_filter_index_iter 🔒

get_packets

Returns a slice of the collected packets opnionally applying the selected filters

get_packets_internal 🔒

get_slice 🔒

merge_filter_type_arrays 🔒