Module grace_lib::general_utils

Helper functions used throughout code.

Modules

SCOPE_ID_COUNTER

Generated by the simple-counter crate.

Functions

c_int

c_union

Take the union of two hashsets, cloning elements from both and consuming neither set.

extend_map

get_next_grad

Generate a globally unique gradual type ID.

get_next_id

get_next_module_id

Generate a globally unique module ID.

get_next_scope_id

Generate a globally unique scope ID.

get_next_var

Generate a globally unique variable ID.

join

Append vec b to vec a and return the result, consuming both.

m_bt_union

Take the union of two btreesets, moving elements out of the second. Both original sets are consumed.

m_union

Take the union of two hashsets, moving elements out of the second. Both original sets are consumed.

mb_union

Take the union of two BTreeSets, consuming both.

unzip_vec

Unzip a vector of two tuples into a tuple of vectors.

vec_c_int

Take the intersection of two vectors, without consuming either.

vec_c_union

Take the union of two vectors, without consuming either.

vec_subset