Struct genetic_planner::genetic_planner::PlannerConfiguration

pub struct PlannerConfiguration {
    pub max_actions: usize,
    pub population_size: usize,
    pub elitism_size: usize,
    pub tournmant_size: usize,
    pub uniform_rate: f32,
    pub mutation_rate: f32,
    pub threadpool_size: usize,
}

Contains the configuration of the Planner

Fields
max_actions: usize

Max number of actions

population_size: usize

Number of the Individual in the Population

elitism_size: usize

Number of Individual to copy in the next generation

tournmant_size: usize

Size of the set used to select the parentof the offsprings

uniform_rate: f32

Parameter used by the crossover function

mutation_rate: f32

Parameter used by the mutate function

threadpool_size: usize

Number of thread used in the evolve function