Module vessels::graphics::canvas

Provides bindings to 2D vector graphic rendering functionality.

Structs

Content
Renderable content.

Enums

Rasterizable
A type that can be rasterized.

Traits

ActiveCanvas
An active canvas.

Canvas
Provides 2-dimensional euclidean rendering capabilities.

CanvasContext
An aggregated context with bound graphics.

Frame
An isolated rendering context.

InactiveCanvas
An inactive canvas.

InteractiveCanvas
A graphics context that can provide interaction and windowing.

Object
Represents content optimized and cached for rendering.

Rasterizer
Provides an interface for the rasterization of content.

Ticker
A type that permits the binding of tick handlers.

Functions

new
Initializes a new graphics context.