Struct interval_set::interval_set::Interval

pub struct Interval(_, _);

Struct Interval containing two values representing the limit of the interval.

The Interval is incluse which means that Interval(0, 10) is [0, 10]. The value 0 is supposed to be equals or greater than the second value.

Methods

impl Interval

fn new(begin: u32, end: u32) -> Interval

fn whole() -> Interval

Return the maximum interval possible (with u32 var)

fn range_size(&self) -> u32

Because the trait Order is needed to sort the IntervalSet I dont what to change the native order. This function coud be considered as the len of the interval.

fn as_tuple(&self) -> (u32, u32)

Simply return an equivalent interval as tuple.

fn get_inf(&self) -> u32

I am not sure about those two function, maybe set the field as public could be a better idea...

fn get_sup(&self) -> u32

fn is_valid(&self) -> bool

Utility function check if the interval is valid.

Examples

The following intervals are valids:

use interval_set::Interval;
Interval::new(0, 0);
Interval::new(10, 100);

The following intervals ae not valid:

use interval_set::Interval;
Interval::new(10, 0);