Struct riscv_sandbox::isa::Instruction

pub struct Instruction(pub u32);

Base structure of an instruction in the RV32I format (just an unsigned 32bits int)

For more information, see the RISC-V reference in the repository

Methods

impl Instruction

pub fn r(opcode: u8, rd: u8, rs1: u8, rs2: u8, funct: u16) -> Instruction

Creates a R type RV32I instruction. These are used for operations with only register operands (e.g. add r1 r1 r2).

Arguments

opcode - The opcode

rd - The destination register

rs1 - The first operand register

rs2 - The second operand register

funct - A 10bits number extending the opcode

pub fn i(opcode: u8, rd: u8, rs1: u8, imm: i32, funct: u8) -> Instruction

Creates a I type RV32I instruction. These are used for operations with a register and an immediate (e.g. addi r1 r1 128).

Arguments

opcode - The opcode

rd - The destination register

rs1 - The register operand

imm - The immediate operand on 12bits

funct - The function to perform (extension of the opcode, on 3bits)

pub fn s(opcode: u8, rs1: u8, rs2: u8, imm: i32, funct: u8) -> Instruction

Creates a S type RV32I instruction. These are used for operations with 2 register operands and an immediate, but no destination register (e.g. stw r1 r2 10).

Arguments

opcode - The opcode

rs1 - First register operand

rs2 - Second register operand

imm - Immediate operand on 12bits representing bits [11:0]

funct - A 3bits extension of the opcode

pub fn b(opcode: u8, rs1: u8, rs2: u8, imm: i32, funct: u8) -> Instruction

Creates a B type RV32I instruction. These are S type instructions with a different immediate layout (the immediate on B represent bits [12:1])

pub fn u(opcode: u8, rd: u8, imm: i32) -> Instruction

Creates a U type RV32I instruction. These are used for operations with only an immediate operand. As it carries fewer information than other instructions, U type instructions have more space for their immediate.

Arguments

opcode - The opcode

rd - The destination register of the operation

imm - bits [31:12] of the 32bits immediate value

pub fn j(opcode: u8, rd: u8, imm: i32) -> Instruction

Creates a J type RV32I instruction. These are U type instructions with a different immediate layout (the immdiate on J represent bits [19:0])

pub fn get_opcode(&self) -> u8

pub fn set_opcode(&mut self, opcode: u8)

pub fn get_rd(&self) -> u8

pub fn set_rd(&mut self, rd: u8)

pub fn get_rs1(&self) -> u8

pub fn set_rs1(&mut self, rs1: u8)

pub fn get_rs2(&self) -> u8

pub fn set_rs2(&mut self, rs2: u8)

pub fn get_funct3(&self) -> u8

pub fn set_funct3(&mut self, funct3: u8)

pub fn get_funct7(&self) -> u8

pub fn set_funct7(&mut self, funct7: u8)

pub fn get_funct10(&self) -> u16

pub fn set_funct10(&mut self, funct10: u16)

pub fn get_imm_i(&self) -> i32

pub fn set_imm_i(&mut self, imm: i32)

pub fn get_imm_s(&self) -> i32

pub fn set_imm_s(&mut self, imm: i32)

pub fn get_imm_u(&self) -> i32

pub fn set_imm_u(&mut self, imm: i32)

pub fn get_imm_b(&self) -> i32

pub fn set_imm_b(&mut self, imm: i32)

pub fn get_imm_j(&self) -> i32

pub fn set_imm_j(&mut self, imm: i32)

pub fn get_type(&self) -> Type

pub fn get_opcode_enum(&self) -> OpCode