kasm_ableton_rust

Module kasm_rulez

Structs

NoteData

Rulez

RulezResult

Enums

MidiEventType

Kasm Rulez - Main interface for instrument realism rules

RuleOutcome

RulezType

Statics

ACTIVE_RULES

Traits

InstrumentRule

Functions

add_rule

Add a rule to the registry (appended at the end)

apply_all_rules

apply_rules_chain

Apply all active rules in order to a set of notes, returning actionable results

apply_rules_to_note

Apply all active rules in order to a single note (convenience)

can_send_note

Returns true if a note can be added according to current rules (polyphony, velocity, hand drum realism, etc.)

clear_rules

Remove all rules from the registry

filter_or_modify_cc

Filter or modify CC data. Return None to block, Some(modified) to allow/modify.

filter_or_modify_note

Filter or modify Note data. Return None to block, Some(modified) to allow/modify.

kasm_config

Configure rules from config struct

remove_rule

Remove all rules of a given type from the registry