Struct keychain_services::keychain::item::Query

pub struct Query(_);

Query builder for locating particular keychain items.

For more information, see "Search Attribute Keys and Values": https://developer.apple.com/documentation/security/keychain_services/keychain_items/search_attribute_keys_and_values

Methods

impl Query

pub fn new() -> Self

Create a new keychain item query builder

pub fn application_label<L: Into<AttrApplicationLabel>>(self, label: L) -> Self

Query for keychain items with the provided SecAttrApplicationLabel (not to be confused with a SecAttrLabel), i.e. the hash/fingerprint of a public key in the keychain.

Both the private and public key in a keypair have a SecAttrApplicationLabel set to the public key's fingerprint.

Wrapper for the kSecAttrApplicationLabel attribute key. See: https://developer.apple.com/documentation/security/ksecattrlabel

pub fn application_tag<T>(self, tag: T) -> Self where
T: Into<AttrApplicationTag>,

Query for keychain items with the provided SecAttrApplicationTag.

Wrapper for the kSecAttrApplicationTag attribute key. See: https://developer.apple.com/documentation/security/ksecattrapplicationtag

pub fn key_class(self, key_class: AttrKeyClass) -> Self

Query for keys with the given SecAttrKeyClass.

Wrapper for the kSecAttrKeyClass attribute key. See: https://developer.apple.com/documentation/security/ksecattrkeyclass

pub fn key_type(self, key_type: AttrKeyType) -> Self

Query for keys with the given SecAttrKeyType.

Wrapper for the kSecAttrKeyType attribute key. See: https://developer.apple.com/documentation/security/ksecattrkeytype

pub fn label<L: Into<AttrLabel>>(self, label: L) -> Self

Query for a particular (human-meaningful) label on keys

Wrapper for the kSecAttrLabel attribute key. See: https://developer.apple.com/documentation/security/ksecattrlabel

pub fn permanent(self, value: bool) -> Self

Query for keys which are or not permanent members of the default keychain.

Wrapper for the kSecAttrIsPermanent attribute key. See: https://developer.apple.com/documentation/security/ksecattrispermanent

pub fn synchronizable(self, value: bool) -> Self

Query for keys which are or are not synchronizable.

Wrapper for the kSecAttrSynchronizable attribute key. See: https://developer.apple.com/documentation/security/ksecattrsynchronizable

pub fn token_id(self, value: AttrTokenId) -> Self

Query for keys stored in an external token i.e. the Secure Enclave Processor (SEP).

Wrapper for the kSecAttrTokenID attribute key. See: https://developer.apple.com/documentation/security/ksecattrtokenid

pub fn use_operation_prompt(self, value: &str) -> Self

Prompt the user with the given custom message when using keys returned from this query.

Wrapper for the kSecUseOperationPrompt. See: https://developer.apple.com/documentation/security/ksecuseoperationprompt