Module lg_webos_client::lg_command::request_commands::web_os_services

WebOs Services Command Module

This modules is related to “webos.service” Uniform Resource. Currently it’s possible to get the WebOs mouse Uri, and maybe create some keyboard application using InsertText, DeleteCharacters``and SendEnterKey`, But these three commands only work if a virtual keyboard is showed

Consider to Not use GetPointerInputSocketUri, use the WebOsClient with SendPointerCommandRequest instead

maybe in the future will have a keyboard abstraction.

Structs

DeleteCharacters

To delete text in virtual keyboard, you must set the number of chars to be deleted.

GetCurrentServicesInformationList

GetPointerInputSocketUri

InsertText

Insert Text in virtual keyboard

SendEnterKey

Is the same of clicking in enter button in the virtual keyborad

SetDisplay3D

This command wasn’t tested