Module rust_action_heroes::lib

Honestly lib is an anti-pattern, this shouldn't be here...

Enums

TransformedInputEvent
For most gameplay, user interaction is filtered into one of these five inputs.

Functions

get_sprite
Given a sprite sheet handle, and a sprite number, return that sprite.

load_sprite_sheet
I think this was basically copy-pasted from the Amehtyst docs.

Type Definitions

Int
For when you're too lazy to decide/remember if your whole numbers are i16 or u8, use Int.