Module rustracer::material

Material module.

Provides:

Different pigments that implement GetColor trait ;

Different BRDF that implement both Eval and ScatterRay trait;

A Material thanks to pigments and BRDF.

Structs

CheckeredPigment

A checkered pigment.

DiffuseBRDF

A class representing an ideal diffuse BRDF (also called “Lambertian”).

ImagePigment

A textured pigment.

Material

A material with a particular pigment and BRDF.

SpecularBRDF

A class representing an ideal mirror BRDF.

UniformPigment

A uniform pigment.

Enums

BRDF

Enum of BRDFs.

Pigment

Enum of pigments.

Traits

Eval

A trait for evaluating a particular BRDF on a parametric surface (u,v).

GetColor

Trait that associates a Color with each point on a parametric surface (u,v).

ScatterRay

A trait for scatter a Ray for a particular BRDF.