Trait rustracer::misc::IsClose

pub trait IsClose<Rhs = Self> {
    fn is_close(&self, other: Self) -> bool;
}

Trait for equivalence between two objects, up to a certain error tolerance.

Primitive elements that compose those objects must implement IsClose trait to derive this trait.

Required Methods

fn is_close(&self, other: Self) -> bool