Module safe_client_gateway::models

Models exposed by this service

Important: Names, Enums and Polymorphism

Every field in the structs that you will see in this documentation is camelCased on serialisation.

Enums are SCREAMING_SNAKE_CASED on serialization and the variant is always put into a type json field for polymorphic cases.

Modules

commons

service