Struct rustracer::normal::Normal

pub struct Normal {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

3D Normal struct.

Note: a 3D normal is a 3D vector with norm equal to 1, but it acts differently when 3D transformation is applied.
So is better to create 2 two different objects (structs) despite similarity and doubling the code.

Fields
x: f32

x component.

y: f32

y component.

z: f32

z component.

Implementations

impl Normal

pub fn neg(&self) -> Self

Return the reversed normal.

pub fn dot(self, other: Normal) -> f32

Compute the dot product between two normals.

pub fn squared_norm(self) -> f32

Return the squared norm (Euclidean length) of a normal.

This is faster than norm if you just need the squared norm.

pub fn norm(self) -> f32

Return the norm (Euclidean length) of a normal.

pub fn normalize(self) -> Self

Modify the normal’s norm so that it becomes equal to 1.

And return the normalized normal.