Function pallet_dex::pallet::dispatchables::mint

pub fn mint<T: Config>(
    asset_a: AssetIdOf<T>,
    asset_b: AssetIdOf<T>,
    amount_a: AssetBalanceOf<T>,
    amount_b: AssetBalanceOf<T>
)

The mint function allows a user to add liquidity to a liquidity pool. Given two assets and their amounts, it either creates a new liquidity pool if it does not already exist for these two assets or adds the provided liquidity to an existing pool. The user will receive LP tokens in return.

Arguments

origin - The origin caller of this function. This should be signed by the user adding liquidity.

asset_a - The identifier for the first type of asset that the user wants to provide.

asset_b - The identifier for the second type of asset that the user wants to provide.

amount_a - The amount of asset_a that the user is providing.

amount_b - The amount of asset_b that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If asset_a and asset_b are the same.

If amount_a or amount_b is 0 or less.

If creating a new liquidity pool would exceed the maximum number of allowed assets (AssetLimitReached).

If adding liquidity to the pool fails for any reason due to arithmetic overflows or underflows

Events

If the function succeeds, it triggers two events:

LiquidityPoolCreated(asset_a, asset_b) if a new liquidity pool was created.

LiquidityAdded(asset_a, asset_b, amount_a, amount_b) after the liquidity has been successfully added.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::mint.