Struct pallet_dex::pallet::Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn mint( origin: OriginFor<T>, asset_a: AssetIdOf<T>, asset_b: AssetIdOf<T>, amount_a: AssetBalanceOf<T>, amount_b: AssetBalanceOf<T> ) -> DispatchResult

The mint function allows a user to add liquidity to a liquidity pool. Given two assets and their amounts, it either creates a new liquidity pool if it does not already exist for these two assets or adds the provided liquidity to an existing pool. The user will receive LP tokens in return.

Arguments

origin - The origin caller of this function. This should be signed by the user adding liquidity.

asset_a - The identifier for the first type of asset that the user wants to provide.

asset_b - The identifier for the second type of asset that the user wants to provide.

amount_a - The amount of asset_a that the user is providing.

amount_b - The amount of asset_b that the user is providing.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If asset_a and asset_b are the same.

If amount_a or amount_b is 0 or less.

If creating a new liquidity pool would exceed the maximum number of allowed assets (AssetLimitReached).

If adding liquidity to the pool fails for any reason due to arithmetic overflows or underflows

Events

If the function succeeds, it triggers two events:

LiquidityPoolCreated(asset_a, asset_b) if a new liquidity pool was created.

LiquidityAdded(asset_a, asset_b, amount_a, amount_b) after the liquidity has been successfully added.

pub fn burn( origin: OriginFor<T>, asset_a: AssetIdOf<T>, asset_b: AssetIdOf<T>, token_amount: AssetBalanceOf<T> ) -> DispatchResult

The burn function allows a user to remove liquidity from a specified liquidity pool. This is done by burning LP tokens and receiving the respective underlying assets in return.

Arguments

origin - The origin caller of this function. This should be signed by the user removing liquidity.

asset_a - The identifier for the first type of asset in the liquidity pool.

asset_b - The identifier for the second type of asset in the liquidity pool.

token_amount - The amount of liquidity the user wants to remove. This is denominated in LP tokens.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If token_amount is 0 or if it’s more than the LP Token balance of the caller

If asset_a and asset_b are the same.

If the liquidity pool for the given asset pair does not exist.

If removing the liquidity from the pool fails for any reason due to arithmetic overflow or underflow

Events

If the function succeeds, it triggers a LiquidityRemoved(asset_a, asset_b, token_amount) event.

pub fn swap( origin: OriginFor<T>, asset_in: AssetIdOf<T>, asset_out: AssetIdOf<T>, amount_in: AssetBalanceOf<T> ) -> DispatchResult

The swap function allows a user to exchange one type of token for another within a specific liquidity pool.

Arguments

origin - The origin caller of this function. This should be signed by the user performing the swap.

asset_in - The identifier for the type of asset that the user wants to swap from.

asset_out - The identifier for the type of asset that the user wants to swap to.

amount_in - The amount of asset_in that the user wants to swap.

Errors

This function will return an error in the following scenarios:

If the origin is not signed (i.e., the function was not called by a user).

If the provided assets do not exist.

If amount_in is 0 or less.

If asset_in and asset_out are the same.

If the liquidity pool for the given pair of assets does not exist.

If the swap operation fails for any reason due to arithmetic error

Events

If the function succeeds, it triggers a Swapped(asset_a, asset_b, amount_in) event.

impl<T: Config> Pallet<T>

pub fn asset_counter() -> AssetIdOf<T>

An auto-generated getter for AssetCounter.