c4_rust_ghiyathi
Module parser

Functions

expect_token ğŸ”’

parses a function call from the token stream

parse

parses a sequence of tokens into an AST

parse_add_sub ğŸ”’

parses an addition or subtraction expression from the token stream

parse_assignment ğŸ”’

parses an assignment statement from the token stream

parse_block ğŸ”’

parses a block of statements enclosed in braces

parse_cmp ğŸ”’

parses a comparison expression from the token stream

parse_declaration ğŸ”’

parses a variable declaration from the token stream

parse_expr ğŸ”’

parses an expression from the token stream

parse_if ğŸ”’

parses an if statement from the token stream

parse_mul_div ğŸ”’

parses a multiplication or division expression from the token stream

parse_primary ğŸ”’

parses a primary expression from the token stream

parse_stmt ğŸ”’

parses an individual statement from the token stream

parse_while ğŸ”’

parses a while loop from the token stream