Module grace_lib::parser::parser_utils::iresult_helpers

Functions

chain

check_data

Check just the data of the result of a parser that includes an update. Skips the containing node and the update.

check_data_and_leftover

check_data_no_update

Check just the data of the result of a parser. Skips the containing node.

check_failed

check_match

Check that the parser applied to the input returns the expect result.

check_match_and_leftover

check_match_no_update

fmap_iresult

Map the contents of an IResult. Rust functors plox

fmap_node

Map data and wrap the result in a Node.

fmap_nodeu

Map data and update with a function that may produce an additional update, and wrap the result in a Node.

fmap_pass

Map data and update with a function that does not produce an additional update, and wrap the result in a Node.

fmap_update

output

simple_check_failed

unwrap_and_check_error

wrap_err