Module grace_lib::parser::parser_utils::tokens

Functions

ADDASN

ALPHA

Recognize a sequence of (ASCII) alphabetic characters.

ALPHANUM

Recognize a sequence of alphanumeric characters.

AND

ARROW

AS

BAND

BANDASN

BORASN

BREAK

BXOR

BXORASN

CLOSE_BRACE

CLOSE_BRACKET

CLOSE_PAREN

COLON

COMMA

CONTINUE

DEQUAL

DIGIT

Recognize a non-empty sequence of decimal digits.

DIGIT0

Recognize a sequence of decimal digits.

DIV

DIVASN

DOT

ELIF

ELSE

EMPTY

Recognize an empty input.

END_OF_INPUT

Recognize an empty input or an end of file

EQUALS

EXCEPT

EXP

EXPASN

FINALLY

FN

FOR

GEQUAL

IDENTIFIER

Parser to an Identifier AST.

IDENT_CHAR

Recognize a sequence of valid internal identifier characters.

IF

IMPL

IMPORT

IN

LANGLE

LEQUAL

LET

LSHASN

LSHIFT

MATCH

MINUS

MOD

MODASN

MULASN

NEG

Parser for the negative unary operator. Checks that it’s not immediately followed by a digit.

NEQUAL

NEWLINE

NOT

NUM_START

OPEN_BRACE

OPEN_BRACKET

OPEN_PAREN

OR

PASS

PLUS

RANGLE

RESERVED

true if a key

RETURN

RSHASN

RSHIFT

SELF

SIGN

STAR

STRING_CHAR

Recognize a character in a string, or an escape sequence.

STRUCT

SUBASN

TARROW

TILDE

TRAIT

TRY

UNDERSCORE

VALID_NUM_FOLLOW

VBAR

WHILE

XOR

YIELD