Struct grace_lib::parser::position_tracker::PosStr

pub struct PosStr<'a> {
    pub offset: usize,
    pub line: u32,
    pub column: u32,
    pub slice: &'a [u8],
}

Fields
offset: usize

The offset represents the position of the slice relatively to the input of the parser. It starts at offset 0.

line: u32

The line number of the slice relatively to the input of the parser. It starts at line 1.

column: u32

The column number of the slice relatively to the input of the parser. It starts at column 1.

slice: &'a [u8]

The slice that is spanned.

Implementations

impl<'a> PosStr<'a>

pub fn new(input: &'a [u8]) -> Self

pub fn new_at(input: &'a [u8], offset: usize, line: u32, column: u32) -> Self

pub fn empty() -> Self

pub fn as_slice(&self) -> &'a [u8]