radar_server::plugins::radar_packet

Trait Hdf5Object

pub trait Hdf5Object {
    // Required methods
    fn to_hdf5(&self, file: &File) -> Result<()>;
    fn from_hdf5(group: &Group, time: SystemTime) -> Result<Self>
       where Self: Sized;
}

An Hdf5Object can be stored and retrieved from an HDF5 file.

Required Methods

fn to_hdf5(&self, file: &File) -> Result<()>

Stores the object in the specified file.

The object is stored in a group corresponding to its epoch time, and a subgroup corresponding to its subsecond millisecond time.

fn from_hdf5(group: &Group, time: SystemTime) -> Result<Self>
where Self: Sized,

Retrieves an object from the specified file and group.

Only retrieves an object from a group or fails, does not search for a specific object.