citadel_sdk
Module prefabs

A list of prefabricated kernels designed for common use cases. If a greater degree of control is required for an application, a custom implementation of NetKernel is desirable

Modules

client

Kernels for clients

server

Kernels for servers

Structs

ClientServerRemote

A limited version of the NodeRemote designed to only allow shutdown calls to the protocol as well as several other functions

Functions

get_socket_addr