citadel_sdk::prefabs::client
Module peer_connection

A kernel that assists in allowing multiple possible peer-to-peer connections

Structs

AddedPeer

FileTransferHandleRx

PeerConnectionKernel

After establishing a connection to the central node, this kernel begins connecting to the desired peer(s)

PeerConnectionSetupAggregator

Allows easy aggregation of UserIdentifier’s and custom settings for the connection request