citadel_sdk::prefabs
Module server

Kernels for servers

Modules

accept_file_transfer_kernel

A kernel that accepts all inbound file transfer requests for basic file transfers AND RE-VFS transfers

client_connect_listener

A kernel that reacts to new channels created, allowing communication with new clients. Useful for when a server needs to send messages to clients

empty

A non-reactive kernel that does no additional processing on top of the protocol

internal_service

For internal services (e.g., a Hyper webserver)