Module citadel_sdk::prefabs::server::client_connect_listener

A kernel that reacts to new channels created, allowing communication with new clients. Useful for when a server needs to send messages to clients

Structs

ClientConnectListenerKernel

A kernel that executes a user-provided function each time a client makes a connection