layers
Module prelude

Re-exports

pub use super::engine::task::AnimationFuture;

pub use super::engine::task::TransitionFuture;

pub use super::engine::AnimationRef;

pub use super::engine::Engine;

pub use super::engine::NodeRef;

pub use super::engine::TransactionRef;

pub use super::types::BlendMode;

pub use super::types::BorderRadius;

pub use super::types::BorderStyle;

pub use super::types::Color;

pub use super::types::PaintColor;

pub use super::types::Point;

pub use super::types::Rectangle;

pub use super::view::BuildLayerTree;

pub use super::view::LayerTree;

pub use super::view::LayerTreeBuilder;

pub use super::view::RenderLayerTree;

pub use super::view::View;

pub use super::engine::animation::*;

Modules

taffy

Structs

ContentDrawError

ContentDrawFunction

Layer

Matrix

PointerHandlerFunction

Scene

Enums

Shape

Shape defines the visual boundary of a layer.

Traits

DrawScene

Effect

Interpolate

Functions

draw_scene

Draw the scene to the given skia::Canvas

render_node_tree

Render a node and his subtree in the provided canvas taking care of the image caching if needed

Type Aliases

Image

Image describes a two dimensional array of pixels to draw. The pixels may be decoded in a raster bitmap, encoded in a Picture or compressed data stream, or located in GPU memory as a GPU texture.