Struct citadel_sdk::prelude::NodeBuilder

pub struct NodeBuilder { /* private fields */ }

Used to construct a running client/peer or server instance

Implementations

impl NodeBuilder

pub fn build<'a, 'b: 'a, K: NetKernel + 'b>(
&'a mut self,
kernel: K
) -> Result<NodeFuture<'b, K>, NodeBuilderError>

Returns a future that represents the both the protocol and kernel execution

pub fn with_node_type(&mut self, node_type: NodeType) -> &mut Self

Defines the node type. By default, Peer is used. If a server is desired, a bind address is expected

use std::net::SocketAddr;
use std::str::FromStr;
use citadel_sdk::prelude::NodeBuilder;
use citadel_proto::prelude::NodeType;

NodeBuilder::default().with_node_type(NodeType::Server(SocketAddr::from_str("0.0.0.0:25021").unwrap()));

pub fn with_backend(&mut self, backend_type: BackendType) -> &mut Self

Sets the backend used to synchronize client account information. By default, uses the filesystem. When the enterprise feature is set, a SQL database (MySQL, PostgreSQL, SQLite) is available. Using a single SQL cluster can be used in combination with a cluster of load-balancing running [‘NetKernel’]’s on different IPs to construct scaled applications

pub fn with_kernel_executor_settings(
&mut self,
kernel_executor_settings: KernelExecutorSettings
) -> &mut Self

Sets the desired settings for the KernelExecutor

pub fn with_server_argon_settings(
&mut self,
settings: ArgonDefaultServerSettings
) -> &mut Self

Attaches custom Argon settings for password hashing at the server

pub fn with_server_misc_settings(
&mut self,
misc_settings: ServerMiscSettings
) -> &mut Self

Attaches miscellaneous server settings (e.g., passwordless mode)

pub fn with_underlying_protocol(
&mut self,
proto: ServerUnderlyingProtocol
) -> &mut Self

Sets the underlying protocol for the server Default: TLS transport w/ self-signed cert

pub async fn with_native_certs(&mut self) -> Result<&mut Self, NodeBuilderError>

Loads the accepted cert chain stored by the local operating system If a custom set of certs is required, run Self::with_custom_certs This is the default if no [RustlsClientConfig] is specified

pub fn with_insecure_skip_cert_verification(&mut self) -> &mut Self

The client will skip unconditionally server certificate verification This is not recommended

pub fn with_custom_certs<T: AsRef<[u8]>>(
&mut self,
custom_certs: &[T]
) -> Result<&mut Self, NodeBuilderError>

Loads a custom list of certs into the acceptable certificate list. Connections that present server certificates that are outside of this list during the handshake process are refused

pub async fn with_pem_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<&mut Self, NodeBuilderError>

The file should be a DER formatted certificate