layers::prelude::taffy
Struct Point

pub struct Point<T> {
    pub x: T,
    pub y: T,
}

A 2-dimensional coordinate.

When used in association with a Rect, represents the top-left corner.

Fields
x: T

The x-coordinate

y: T

The y-coordinate

Implementations

impl Point<f32>

pub const ZERO: Point<f32>

A Point with values (0,0), representing the origin

impl Point<Option<f32>>

pub const NONE: Point<Option<f32>>

A Point with values (None, None)

impl<T> Point<T>

pub fn map<R, F>(self, f: F) -> Point<R>
where F: Fn(T) -> R,

Applies the function f to both the x and y

This is used to transform a Point<T> into a Point<R>.

pub fn get(self, axis: AbstractAxis) -> T

Gets the extent of the specified layout axis Whether this is the width or height depends on the GridAxis provided

pub fn transpose(self) -> Point<T>

Swap x and y components

pub fn set(&mut self, axis: AbstractAxis, value: T)

Sets the extent of the specified layout axis Whether this is the width or height depends on the GridAxis provided

impl<T> Point<T>
where T: TaffyZero,

pub const fn zero() -> Point<T>

Returns a Point where both the x and y values are the zero value of the contained type (e.g. 0.0, Some(0.0), or Dimension::Length(0.0))

impl<T> Point<T>
where T: TaffyAuto,

pub const fn auto() -> Point<T>

Returns a Point where both the x and y values are the auto value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Point<T>
where T: TaffyMinContent,

pub const fn min_content() -> Point<T>

Returns a Point where both the x and y values are the min_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Point<T>
where T: TaffyMaxContent,

pub const fn max_content() -> Point<T>

Returns a Point where both the x and y values are the max_content value of the contained type (e.g. Dimension::Auto or LengthPercentageAuto::Auto)

impl<T> Point<T>
where T: TaffyFitContent,

pub fn fit_content(argument: LengthPercentage) -> Point<T>

Returns a Point with x and y set to the same fit-content(…) value with the given argument.

impl<T> Point<T>
where T: FromLength,

pub fn length<Input>(value: Input) -> Point<T>
where Input: Into<f32> + Copy,

Returns a Point where x and y values are the same given absolute length

impl<T> Point<T>
where T: FromPercent,

pub fn percent<Input>(percent_value: Input) -> Point<T>
where Input: Into<f32> + Copy,

Returns a Point where both the x and y values are the constant percent value of the contained type (e.g. 2.1, Some(2.1), or Dimension::Length(2.1))