Trait elma_builder::process::Process

pub trait Process {
    fn period(&self) -> Duration;

    fn previous_update(&self) -> Duration;

    fn last_update(&self) -> Duration;

    fn start_time(&self) -> SystemTime;

    fn name(&self) -> &String;

    fn num_updates(&self) -> i64;

    fn status(&self) -> &StatusEnum;

    fn set_status(&mut self, status: StatusEnum);

    fn set_start_time(&mut self, st: SystemTime);

    fn set_prev_update(&mut self, pu: Duration);

    fn set_last_update(&mut self, lu: Duration);

    fn set_num_update(&mut self, nu: i64);

    fn set_period(&mut self, per: Duration);

    fn sum(&self) -> f64 { ... }

    fn _update(&mut self, _c: &mut Vec<Box<Channel>>, _elapsed: Duration) { ... }

    fn milli_time(&self) -> u64 { ... }

    fn status_type_map(&self) -> String { ... }

    fn delta(&self) -> u64 { ... }

    fn _init(&mut self) { ... }

    fn _stop(&mut self) { ... }

    fn _start(&mut self, _elapsed: Duration) { ... }
}

Remarks

Note that the base implementation a process in Rust is through a trait. This is done to enforce certain rules for all instantiations of a process. Ensure that all processes created utilize the process trait.

Functions given here are a mix of just a signature or may include the functionality. All functions in this trait are available to use in the implementation of the trait, but one also has the opportunity to shadow any given function given below in the implementation as long as it doesn't conflict with the fn signature. This means all given functions in this trait are the equivalent of a virtual function in C++

Of important note, if there is only a signature, that function needs expansion in the implementer. If the functionality is called out below, then it doesn't need reimplementing in the implementation unless you want to overshadow it. We have thus provided definete rules for all processes.

Required methods

fn period(&self) -> Duration

fn previous_update(&self) -> Duration

fn last_update(&self) -> Duration

fn start_time(&self) -> SystemTime

fn name(&self) -> &String

fn num_updates(&self) -> i64

fn status(&self) -> &StatusEnum

fn set_status(&mut self, status: StatusEnum)

fn set_start_time(&mut self, st: SystemTime)

fn set_prev_update(&mut self, pu: Duration)

fn set_last_update(&mut self, lu: Duration)

fn set_num_update(&mut self, nu: i64)

fn set_period(&mut self, per: Duration)

Loading content...
Provided methods

fn sum(&self) -> f64

fn _update(&mut self, _c: &mut Vec<Box<Channel>>, _elapsed: Duration)

fn milli_time(&self) -> u64

fn status_type_map(&self) -> String

fn delta(&self) -> u64

fn _init(&mut self)

fn _stop(&mut self)

fn _start(&mut self, _elapsed: Duration)

Loading content...