Struct test_data_generation::profile::fact::Fact

pub struct Fact {
    pub key: char,
    pub prior_key: Option<char>,
    pub next_key: Option<char>,
    pub pattern_placeholder: char,
    pub starts_with: u32,
    pub ends_with: u32,
    pub index_offset: u32,
}

Represents a Fact for a character in a sample data entity that has been analyzed

Fields
key: char

the char that the fact defines (.e.g: 'a', '1', '%', etc.)

prior_key: Option<char>

the char that appears before (-1) the key in the entity

next_key: Option<char>

the char that appears after (+1) the key in the entity

pattern_placeholder: char

the PatternPlaceholder symbol that represents the type of key

starts_with: u32

indicates if the key is the first char in the entity (0=no, 1=yes)

ends_with: u32

indicates if the key is the last char in the entity (0=no, 1=yes)

index_offset: u32

indicates the number of positions from the index zero (where the char is located in the entity from the first position)

Methods

impl Fact

fn new(k: char, pp: char, sw: u32, ew: u32, idx_off: u32) -> Fact

Constructs a new Fact

Example

extern crate test_data_generation;

use test_data_generation::profile::fact::Fact;
    
fn main() {
    //fact created for the character 'r' in the string "word"
    let mut fact =  Fact::new('r','c',0,0,2);
}

fn set_next_key(&mut self, nk: char)

This function sets the next key attribute to the specified char.

Arguments

nk: char - The character that represents the next character in the entity

Example

extern crate test_data_generation;

use test_data_generation::profile::fact::Fact;
    
fn main() {
    //fact created for the character 'r' in the string "word"
    let mut fact =  Fact::new('r','c',0,0,2);
    fact.set_next_key('d');
}
 

fn set_prior_key(&mut self, pk: char)

This function sets the prior key attribute to the specified char.

Arguments

pk: char - The character that represents the prior character in the entity

Example

extern crate test_data_generation;

use test_data_generation::profile::fact::Fact;
    
fn main() {
    //fact created for the character 'r' in the string "word"
    let mut fact =  Fact::new('r','c',0,0,2);
    fact.set_prior_key('o');
}