Struct test_data_generation::profile::pattern::Pattern

pub struct Pattern {
    pub size: u32,
    // some fields omitted
}

Represents a symbolic pattern of an entity (String)

Fields
size: u32

The size (length) of the pattern

Methods

impl Pattern

fn new() -> Pattern

Constructs a new Pattern

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern::Pattern;
    
fn main() {
    let mut pattern =  Pattern::new();
}

fn symbolize_char(&self, c: &char) -> char

This function converts a char into a pattern symbol

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern::Pattern;
 
fn main() {
    let pattern =  Pattern::new();
 
    println!("The pattern symbol for 'A' is {:?}", pattern.symbolize_char(&'A'));
    // The pattern symbol for 'A' is V
}

fn analyze(&mut self, entity: &str) -> (String, Vec<Fact>)

This function converts an entity (&str) into a tuplet (String, Vec) String : The symbolic pattern Vec : List of Facts extracted from the entity and related to the symbolic pattern

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern::Pattern;
 
fn main() {
    let mut pattern =  Pattern::new();
 
    assert_eq!(pattern.analyze("Hello World").0, "CvccvSCvccc");
}