Module test_data_generation::profile::pattern_placeholder

The pattern_placeholder module provides functionality to retrieve symbols that are used in defining a pattern.

Here is the list of symbols that identify a type of character: @ = unknown [Unknonw] C = upper case consonant [ConsonantUpper] c = lower case consonant [ConsonantLower] V = upper case vowel [VowelUpper] v = lower case vowel [VowelLower] # = numeric digit [Numeric] ~ = special regex character [RegExSpcChar] S = white space [WhiteSpace] p = punctuation [Punctuation]

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern_placeholder::PatternPlaceholder;
 
fn main() {
    // return the symbol that represents an upper case vowel
    pub fn upper_vowel_symbol() -> char {
        let placeholder =  PatternPlaceholder::new();
        placeholder.get(&"VowelUpper".to_string())
    }

    // confirm that the symbol used to represent an upper case vowel is 'V' 
    assert_eq!(upper_vowel_symbol(), 'V');
}

Structs

PatternPlaceholder
Represents a placeholder for all the symbols used in pattern definitions