Module test_data_generation::profile::profile

The profile module provides functionality to create a profile on a data sample (Strings) based on the following attributes:

symbolic patterns

lengths

pattern probability

Markov chains

Once a profile has been made, data can be generated by calling the _pre_generate()_ and generate() functions, in that order.

Examples

Profiling some sample data...

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    // analyze the dataset
    let mut data_profile =  Profile::new();

    // analyze the dataset
    data_profile.analyze("Smith, John");
    data_profile.analyze("Doe, John"); 
    data_profile.analyze("Dale, Danny"); 
    data_profile.analyze("Rickets, Ronney");

    // confirm 4 data samples were analyzed         
    assert_eq!(data_profile.patterns.len(), 4);
}

Generating a person's name from some sample data...

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut data_profile =  Profile::new();

    // analyze the dataset
    data_profile.analyze("Smith, John");
    data_profile.analyze("Doe, John"); 
    data_profile.analyze("Dale, Danny"); 
    data_profile.analyze("Rickets, Ronney"); 
    data_profile.analyze("Taylor, Mike"); 
    data_profile.analyze("Bates, Betty"); 

    // prepare the generator
    data_profile.pre_generate();
                
    // generate some data
    println!("The generated name is {:?}", data_profile.generate());
}

Generating a date from some sample data...

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut data_profile =  Profile::new();

    // analyze the dataset
    data_profile.analyze("01/01/2017");
    data_profile.analyze("01/02/2017"); 
    data_profile.analyze("01/03/2017"); 
    data_profile.analyze("01/04/2017"); 
    data_profile.analyze("01/05/2017");  
    data_profile.analyze("01/06/2017");  
    data_profile.analyze("01/07/2017");  
    data_profile.analyze("01/08/2017");  
    data_profile.analyze("01/09/2017"); 
    data_profile.analyze("01/10/2017");  

    // prepare the generator
    data_profile.pre_generate();
                
    // generate some data
    println!("The generated date is {:?}", data_profile.generate());
}

Structs

Profile
Represents a Profile for sample data that has been analyzed and can be used to generate realistic data