Struct test_data_generation::profile::profile::Profile

pub struct Profile {
    pub patterns: BTreeMap<String, u32>,
    pub pattern_total: u32,
    pub pattern_keys: Vec<String>,
    pub pattern_vals: Vec<u32>,
    pub pattern_percentages: Vec<(String, f64)>,
    pub pattern_ranks: Vec<(String, f64)>,
    pub sizes: BTreeMap<u32, u32>,
    pub size_total: u32,
    pub size_ranks: Vec<(u32, f64)>,
    pub processors: u8,
    pub facts: Vec<Vec<Fact>>,
}

Represents a Profile for sample data that has been analyzed and can be used to generate realistic data

Fields
patterns: BTreeMap<String, u32>

A list of symbolic patterns with a distinct count of occurrences

pattern_total: u32

The total number of patterns in the profile

pattern_keys: Vec<String>

A list of symbolic patterns in the profile (used for temporary storage due to lifetime issues)

pattern_vals: Vec<u32>

A list of distinct counts for patterns in the profile (used for temporary storage due to lifetime issues)

pattern_percentages: Vec<(String, f64)>

A list of symbolic patterns with their percent chance of occurrence

pattern_ranks: Vec<(String, f64)>

A list of symbolic patterns with a running total of percent chance of occurrence, in increasing order

sizes: BTreeMap<u32, u32>

A list of pattern lengths with a distinct count of occurrence

size_total: u32

the total number of pattern sizes (lengths) in the profile

size_ranks: Vec<(u32, f64)>

A list of pattern sizes (lengths) with a running total of their percent chance of occurrence, in increasing order

processors: u8

The number of processors used to distribute the work load (multi-thread) while finding Facts to generate data

facts: Vec<Vec<Fact>>

A list of processors (which are lists of Facts) that store all the Facts in the profile

Methods

impl Profile

fn new() -> Profile

Constructs a new Profile

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
    
fn main() {
    let placeholder = Profile::new();
}

fn new_with(p: u8) -> Profile

Constructs a new Profile with a specified number of processors to analyze the data. Each processor shares the load of generating the data based on the Facts it has been assigned to manage.

Arguments

p: u8 - A number that sets the number of processors to start up to manage the Facts. Increasing the number of processors will speed up the generator be ditributing the workload. The recommended number of processors is 1 per 10K data points (e.g.: profiling 20K names should be handled by 2 processors) NOTE: The default number of processors is 4.

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
    
fn main() {
    let processors: u8 = 10;
    let placeholder = Profile::new_with(processors);
}

fn analyze(&mut self, entity: &str)

This function converts an data point (&str) to a pattern and adds it to the profile

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();
    profile.analyze("One");
    profile.analyze("Two"); 
    profile.analyze("Three"); 
    profile.analyze("Four"); 
                
    assert_eq!(profile.patterns.len(), 4);
}

fn apply_facts(&self, pattern: String) -> String

This function generates realistic test data based on the sample data that was analyzed.

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();

    profile.analyze("01/13/2017");
    profile.analyze("11/24/2017"); 
    profile.analyze("08/05/2017");
                
    profile.pre_generate();

    let generated = profile.apply_facts("##p##p####".to_string());

    assert_eq!(generated.len(), 10);
}

fn cum_patternmap(&mut self)

This function calculates the patterns to use by the chance they will occur (as cumulative percentage) in decreasing order

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();

    profile.analyze("Smith, John");
    profile.analyze("O'Brian, Henny"); 
    profile.analyze("Dale, Danny"); 
    profile.analyze("Rickets, Ronnae"); 
    profile.analyze("Richard, Richie");
    profile.analyze("Roberts, Blake");
    profile.analyze("Conways, Sephen");
        
    profile.pre_generate(); 
    let test = [("CvccvccpSCvccvv".to_string(), 28.57142857142857 as f64), ("CcvccpSCvcc".to_string(), 42.857142857142854 as f64), ("CvccvccpSCvccvc".to_string(), 57.14285714285714 as f64), ("CvcvcccpSCcvcv".to_string(), 71.42857142857142 as f64), ("CvcvpSCvccc".to_string(), 85.7142857142857 as f64), ("V@CcvvcpSCvccc".to_string(), 99.99999999999997 as f64)];            
                    
    assert_eq!(profile.pattern_ranks, test);
}

fn cum_sizemap(&mut self)

This function calculates the sizes to use by the chance they will occur (as cumulative percentage) in decreasing order

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();
    profile.analyze("One");
    profile.analyze("Two"); 
    profile.analyze("Three"); 
    profile.analyze("Four");  
    profile.analyze("Five");
    profile.analyze("Six");
                
    profile.cum_sizemap();

    print!("The size ranks are {:?}", profile.size_ranks);
    // The size ranks are [(3, 50), (4, 83.33333333333333), (5, 100)] 
}

fn generate(&mut self) -> String

This function generates realistic test data based on the sampel data that was analyzed.

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();

    profile.analyze("One");
    profile.analyze("Two"); 
    profile.analyze("Three"); 
    profile.analyze("Four");  
    profile.analyze("Five");
                
    profile.pre_generate();

    print!("The test data {:?} was generated.", profile.generate());
}

fn pre_generate(&mut self)

This function prepares the size a pattern accumulated percentages order by percentage increasing

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();
    profile.analyze("One");
    profile.analyze("Two"); 
    profile.analyze("Three"); 
    profile.analyze("Four");  
    profile.analyze("Five");
    profile.analyze("Six");
                
    profile.pre_generate();

    print!("The size ranks are {:?}", profile.size_ranks);
    // The size ranks are [(3, 50), (4, 83.33333333333333), (5, 100)] 
}

fn reset_analyze(&mut self)

This function resets the patterns that the Profile has analyzed. Call this method whenever you wish to "clear" the Profile

Example

extern crate test_data_generation;

use test_data_generation::profile::profile::Profile;
 
fn main() {
    let mut profile =  Profile::new();

    profile.analyze("One");
    profile.analyze("Two"); 
    profile.analyze("Three"); 
    
    let x = profile.patterns.len();

    profile.reset_analyze();

    profile.analyze("Four");
    profile.analyze("Five"); 
    profile.analyze("Six");  
    profile.analyze("Seven"); 
    profile.analyze("Eight"); 
    profile.analyze("Nine"); 
    profile.analyze("Ten");
    
    let y = profile.patterns.len();
                
    assert_eq!(x, 3);
    assert_eq!(y, 5);
}