Struct decapod::reader::Read

pub struct Read { /* private fields */ }

pod5 read information. see https://github.com/nanoporetech/pod5-file-format/blob/master/docs/tables/reads.toml for full pod5-file-format specification information.

Implementations

impl Read

pub fn signal(&self) -> Result<Vec<i16>>

The uncompressed signal for the associated read.

pub fn uuid(&self) -> Uuid

The read id as a uuid.

pub fn read_number(&self) -> u32

The read number.

pub fn start_sample(&self) -> u64

Number of samples in the channel before read was taken.

pub fn median_before(&self) -> f32

Current level before read taken.

pub fn channel(&self) -> u16

1-indexed channel.

pub fn well(&self) -> u8

1-indexed well number.

pub fn pore_type(&self) -> i16

Raw pore type.

pub fn pore_type_string(&self) -> Result<String>

Attempts to get the pore type as a String.

pub fn calibration(&self) -> Calibration<'_>

Calibration and extra calibration data for the read.

pub fn end_reason(&self) -> EndReason

See crate::endreason::EndReason

pub fn end_reason_forced(&self) -> bool

True if the read was forcefully ended.

pub fn run_info_num(&self) -> i16

Which run info number.

pub fn run_info(&self) -> Result<RunInfo>

The runinfo struct from the meta data for the associated read.

pub fn num_minknow_events(&self) -> u64

Number of minknow events that the read contains.

pub fn tracked_scaling_scale(&self) -> f32

Tracked scaling scale.

pub fn tracked_scaling_shift(&self) -> f32

Tracked scaling shift.

pub fn predicted_scaling_scale(&self) -> f32

Predicted scaling scale.

pub fn predicted_scaling_shift(&self) -> f32

Predicted scaling shift.

pub fn num_reads_since_mux_change(&self) -> u32

Number of reads since last mux change.

pub fn time_since_mux_change(&self) -> f32

Time in seconds since last mux change.

pub fn signal_row_count(&self) -> i64

Number of signal row count.

pub fn num_samples(&self) -> u64

Number of signal samples.