Struct decapod::reader::Reader

pub struct Reader { /* private fields */ }

Open pod5 files, directories of pod5 files and iterate over records and reads.

Features

Opening directories consisting of pod5 files requires the recursive feature to be enabled.

Example

use decapod::reader::Reader
use uuid::Uuid
use std::error::Error

fn main() -> Result<(), Box<dyn Error>>
{
    let reader = Reader::from_path("example.pod5", None)?;

    println!("{:?}", &reader.read_ids()?);
    Ok(())
}

Implementations

impl Reader

pub fn from_path<P: AsRef<Path>>( path: P, options: Option<ReaderOptions>, ) -> Result<Reader>

Opens a pod5 file or directory of pod5 files for reading.

Arguments

path - The path to either a pod5 or a directory containing pod5 files.

options - the ReaderOptions object. Pass None to use no options (typical).

pub fn from_vec<P>( paths: Vec<P>, options: Option<ReaderOptions>, ) -> Result<Reader>
where P: AsRef<Path>,

Opens a combinations of both pod5 file paths and directories containing pod5 files.

Arguments

paths - The vector of pod5 file and directory paths.

options - the ReaderOptions object. Pass None to use no options (typical).

pub fn from_iter<P, I>( iter: I, options: Option<ReaderOptions>, ) -> Result<Reader>
where I: IntoIterator<Item = P>, P: AsRef<Path>,

Opens a combinations of both pod5 file paths and directories containing pod5 files.

Arguments

iter - The iterator object consisting of pod5 file and directory paths.

options - the ReaderOptions object. Pass None to use no options (typical).

pub fn count(&self) -> Result<usize>

Returns the total number of read ids from all open pod5 files.

pub fn read_ids(&self) -> Result<Vec<Uuid>>

Returns the uuids for every read id from all open pod5 files.

pub fn info(&self) -> Vec<Result<FileInfo>>

Returns a vector of all the file info structs found within the open pod5 files.

pub fn run_info_iter(&self) -> RunInfoIter<'_>

Obtain the runinfo iterator.

pub fn reads_iter(&self, fetch: Option<Vec<Uuid>>) -> Reads<'_>

Create a Reads iterator for reads found within the open pod5 files.

Arguments

fetch - Specify None to obtain all reads, or a vector of uuids for specific reads of interest

Example

let reader = Reader::from_path("sample.pod5", None)?;
let read_ids = vec![uuid!("002fde30-9e23-4125-9eae-d112c18a81a7")];
for read in reader.reads_iter(Some(read_ids))
{
    let read = read?;
    println!("{}", read.read_id()?);
}

Returns

A reads iterator.

pub fn batch_records_iter( &self, fetch: Option<Vec<Uuid>>, ) -> BatchRecordIter<'_>

Obtain the batch records iterator. Currently only useful for converting a record of reads to a Polars DataFrame.

Arguments

fetch - Specific read ids can be requested in the same way as Reader::reads_iter.