Module wasm_svg_graphics::renderer

Container for the actual renderer object, this includes all logic for adding items to the DOM and for detecting duplication

Structs

Renderer
Container object used to interact with the SVG Object Keeps track of definitions and dom root id

Constants

DEFAULT_VIEWBOX
The value to which the view box of every svg Renderer will be set to by default. The viewbox of any Renderer, can be adjusted with the adjust_viewbox method.