Function wirefish::report::write_report

pub fn write_report(
    output_path: &str, 
    data: &mut HashMap<SourceDestination, PacketExchange>, 
    first_generation: bool
) -> Result<bool, Error>

Appends data to a report file, creates the file if it doesn’t exist

The file and the directory path to it are created if they do not exist. The hashmap is consumed and its content is written in the csv file indicated by the path. If the first_generation attribute it’s true any file corresponding to the provided path will be deleted and a new file will be generated with a header containing the name of the fields.pub fn write_report(output_path: &str, data: &mut HashMap<SourceDestination, PacketExchange>, first_generation: bool) -> Result<bool, io::Error> {