Module citadel_sdk::responses

For easy construction of replies to common message types A list of helpers making the response phase simple and intuitive

Generally, when making a response, the response must be sent outbound through the [NodeRemote] with a custom ticket equivalent to the ticket receieved by the external request in order for the peer to listen for a response. Additionally, [PeerConnectionType] data must be .reverse()’d. Finally, some response types require the manual input of usernames, and as such, this helper library enforces all these requirements

Functions

group_invite

Given a group invite signal, this function sends a response to the server

peer_connect

Given the input_signal from the peer, this function sends a connect response to the target peer

peer_register

Given the input_signal from the peer, this function sends a register response to the target peer