Module vcf_rider::rider

Structs

RiderParameters

Traits

CanScoreSequence
Our vcf_rider main function will receive a Vec and call it for every T on subsequences of the genomes of the samples doing it only for each variant subsequence once. This trait will need to be able to compute a score on a given sequence, represented by a splice of an array of u8 [TODO] starting for a given position (it is guaranteed by the lib that the used position will be given inside the sequence, i.e. sequence.len() - self.get_length() >= 0).

Functions

encode_genotypes

find_overlapping_snps
Function that advances on the VcfReader (Iterator of Mutation) until the first snp that does not overlap with the given window, putting in snps_buffer all the overlapping snps and their number and then the first not overlapping snp.

get_scores

match_indexes

obtain_seq

print_overlapping