Function docs::routes::about::routes::get_master_copies

pub async fn get_master_copies(
    context: RequestContext, 
    chain_id: String
) -> Result<RawJson<String>, ApiError>

/v1/chains/<chain_id>/about/master-copies
Returns a list of MasterCopy

Master Copies

This endpoint returns a list of MasterCopy objects just as documented in the core services

Path

/v1/chains/<chain_id>/about/master-copies where chain_id correspond to the chain id of the desired network

Sample Json

JSON sample

[
  {
    "address": "0x8942595A2dC5181Df0465AF0D7be08c8f23C93af",
    "version": "0.1.0"
  }
]