Function docs::routes::collectibles::routes::get_collectibles_paginated

pub async fn get_collectibles_paginated(
    context: RequestContext,
    chain_id: String,
    safe_address: String,
    cursor: Option<String>,
    trusted: Option<bool>,
    exclude_spam: Option<bool>
) -> Result<Json<Page<Collectible>>, ApiError>

/v2/chains/<chain_id>/safes/<safe_address>/collectibles?<cursor>&<trusted>&<exclude_spam>
Returns collectibles paginated

Collectibles

The collectibles endpoint does not implement any logic in the client-gateway. The response from the core services is cached and then forwarded to the clients.

Path

/v1/chains/<chain_id>/safes/<safe_address>/collectibles?<cursor>&<trusted>&<exclude_spam> : Returns a list of the ERC721 tokens stored in a safe

Query parameters

<cursor> : Pagination parameters (limit and offset) <trusted> : A token is defined as trusted by our core handlers process when adding them. Default value is false <exclude_spam>: A token is defined as spam by our core handlers process when adding them. Default value is true