Function docs::routes::contracts::routes::post_data_decoder

pub async fn post_data_decoder<'e>(
    context: RequestContext, 
    chain_id: String, 
    data_decoder_request: Result<Json<DataDecoderRequest>, Error<'e>>
) -> Result<RawJson<String>, ApiError>

/v1/chains/<chain_id>/data-decoder
Returns DataDecoded

Data Decoder

This endpoint requires the client to send in the body of the request a hexadecimal String containing the data field of a transaction for decoding

The result is of the type DataDecoded

Path

/v1/chains/<chain_id>/data-decoder

Examples

Example request body:

{
  "data": "0x095ea7b3000000000000000000000000ae9844f89d98c150f5e61bfc676d68b4921559900000000000000000000000000000000000000000000000000001c6bf52634000"
}

This results in:

{
  "method": "approve",
  "parameters": [
   {
      "name": "spender",
      "type": "address",
      "value": "0xae9844F89D98c150F5e61bfC676D68b492155990"
    },
    {
      "name": "value",
      "type": "uint256",
      "value": "500000000000000"
    }
  ]
}