Module docs::routes::transactions

Transactions endpoints

As presented by the endpoints in this service, we are taking in the types returned by the transaction service, which to this data are Multisig, Module and Ethereum transaction types.

The types served by the gate way are Transfer, SettingsChange and Custom. Additionally, we treat the Creation transaction as one additional type, as it is meant to be group with the rest of the items in the same UI component in the apps.

Functions

get_transactions
/v1/chains/<chain_id>/transactions/<transaction_id>
Returns TransactionDetails

get_transactions_history
/v1/chains/<chain_id>/safes/<safe_address>/transactions/history?<cursor>&<timezone_offset>&<trusted>
Returns a Page of TransactionListItem

get_transactions_queued
/v1/chains/<chain_id>/safes/<safe_address>/transactions/queued?<cursor>&<timezone_offset>&<trusted>
Returns a Page of TransactionListItem

post_confirmation
/v1/chains/<chain_id>/transactions/<safe_tx_hash>/confirmations
Returns TransactionDetails

post_transaction
/v1/chains/<chain_id>/transactions/<safe_address>/propose
No return value