Function docs::routes::transactions::routes::get_transactions_history

pub async fn get_transactions_history(
    context: Context<'_>, 
    chain_id: String, 
    safe_address: String, 
    cursor: Option<String>, 
    timezone_offset: Option<String>
) -> Result<Json<String>, ApiError>

/v1/chains/<chain_id>/safes/<safe_address>/transactions/history?<cursor>&<timezone_offset>&<trusted>
Returns a Page of TransactionListItem

Transactions History

This endpoint returns all the transactions that have been executed for a given safe. Cancelled transactions will not be shown in this endpoint. Therefore, there is no concept of conflicting nonces for this endpoint, as there could potentially be for queued transactions.

This endpoint does not return any TransactionListItem::Label nor TransactionListItem::ConflictHeader as of the writing of this iteration of this document.

Transaction are aggregated by day and for each day there is a TransactionListItem::DateLabel added. The timestamp returned corresponds to the date only, time fields should be therefore ignored. The dates are returned in UTC, in a later iteration this will be offset by the timezone_offset sent by the clients in the query parameter.

TransactionListItem::Transaction is returned with the same data layout as in the /transactions/queued endpoint.

The structure of the transaction object corresponds to that of a [crate::models::handlers::transactions::summary::TransactionSummary]

Path

GET /v1/chains/<chain_id>/safes/<safe_address>/transactions/history?<cursor>&<timezone_offset>&<trusted>

Query parameters

<safe_address> should be the checksummed address of the safe to be observed.

<cursor> is the desired page of data to be loaded. Values for this parameter can be either Page.next or Page.previous. WARNING: Don’t fiddle with the values of these 2 fields.

<timezone_offset>: Currently ignored by the gateway.

<trusted>: forwarded directly to the core services. Only for debugging purposes clients should not send it (unless they know what they are doing).