Function docs::routes::transactions::routes::post_confirmation

pub async fn post_confirmation<'e>(
    context: RequestContext, 
    chain_id: String, 
    safe_tx_hash: String, 
    tx_confirmation_request: Result<Json<ConfirmationRequest>, Error<'e>>
) -> Result<Json<String>, ApiError>

/v1/chains/<chain_id>/transactions/<safe_tx_hash>/confirmations
Returns TransactionDetails

Transaction Confirmation

This endpoint provides a way for submitting confirmations for clients making use of the safe_tx_hash as part of the path, and the very same safe_tx_hash signed by an owner corresponding to the safe from which the transaction is being sent.

If the confirmation is submitted successfully to the core services, then the local cache for that specific transaction is invalidated and the updated transaction details with the confirmation are returned in the request.

Path

POST /v1/chains/<chain_id>/transactions/<safe_tx_hash>/confirmations

The expected (ConfirmationRequest)crate::routes::transactions::models::requests::ConfirmationRequest body for this request, as well as the returned (TransactionDetails)crate::routes::transactions::models::details::TransactionDetails

Query parameters

No query parameters available for this endpoint.