Struct rest::scf_io::ScfTraceRecord

pub struct ScfTraceRecord {

    pub num_iter: usize,
    pub mixer: String,
    pub num_max_records: usize,
    pub mix_param: f64,
    pub start_diis_cycle: usize,
    pub scf_energy: f64,
    pub energy_records: Vec<f64>,
    pub prev_hamiltonian: Vec<[MatrixUpper<f64>; 2]>,
    pub eigenvectors: [MatrixFull<f64>; 2],
    pub eigenvalues: [Vec<f64>; 2],
    pub density_matrix: [Vec<MatrixFull<f64>>; 2],
    pub target_vector: Vec<[MatrixFull<f64>; 2]>,
    pub error_vector: Vec<Vec<f64>>,

}

Fields
num_iter: usizemixer: Stringnum_max_records: usizemix_param: f64start_diis_cycle: usizescf_energy: f64energy_records: Vec<f64>prev_hamiltonian: Vec<[MatrixUpper<f64>; 2]>eigenvectors: [MatrixFull<f64>; 2]eigenvalues: [Vec<f64>; 2]density_matrix: [Vec<MatrixFull<f64>>; 2]target_vector: Vec<[MatrixFull<f64>; 2]>error_vector: Vec<Vec<f64>>
Implementations

impl ScfTraceRecord

pub fn new(
num_max_records: usize,
mix_param: f64,
mixer: String,
start_diis_cycle: usize
) -> ScfTraceRecord

pub fn initialize(scf: &SCF) -> ScfTraceRecord

pub fn update(&mut self, scf: &SCF)

This subroutine updates:
scf_energy: from previous to the current value
scf_eigenvalues: from previous to the current value
scf_eigenvectors: from previous to the current value
scf_density_matrix: [pre, cur]
num_iter
This subroutine should be called after [scf.check_scf_convergence] and before [self.prepare_next_input]“

pub fn prepare_next_input(&mut self, scf: &mut SCF)

This subroutine prepares the fock matrix for the next step according different mixing algorithm

self.mixer =

“direct”: the output density matrix in the current step n0[out] will be used directly to generate the the input fock matrix of the next step

“linear”: the density matrix used in the next step n1[in] is a mix between the input density matrix in the current step n0[in] and n0[out]
n1[in] = alpha*n0[out] + (1-alpha)*n0[in]
= n0[in] + alpha * Rn0
where alpha the mixing parameter obtained from self.mix_param and Rn0 = n0[out]-n0[in] is the density matrix change in the current step. n1[in] is then be used to generate the input fock matrix of the next step

“diis”: the input fock matrix of the next step f1[in] = sum_{i} c_i*f_i[in], where f_i[in] is the input fock matrix of the ith step and c_i is obtained by the diis altogirhm against the error vector of the commutator (f_i[out]*d_i[out]*s-s*d_i[out]*f_i[out]), where
- f_i[out] is the ith output fock matrix,
- d_i[out] is the ith output density matrix,
- s is the overlap matrix

Ref: P. Pulay, Improved SCF Convergence Acceleration, JCC, 1982, 3:556-560.