Trait rustracer::shape::RayIntersection

pub trait RayIntersection: Debug + Send + Sync {
    fn ray_intersection(&self, ray: Ray) -> Option<HitRecord>;
}

Trait to determine the intersections of an object with a Ray.

This trait is meant to be implemented by geometrical shapes in order to calculate how light rays hit them.

All objects that implement RayIntersection must implement also:

Debug;

Send;

Sync.

Required Methods

fn ray_intersection(&self, ray: Ray) -> Option<HitRecord>