Module rust_action_heroes::state

Loading, Menu, Level

Amethyst States are like clear plastic you can draw on. When you want a new drawing surface, you put a new state on top.

When you put on new state on, you may want to cleanup your current state, or leave it as is.

This game uses a lot of "push" state to add a state to the stack, "pop" state to go back to the previous state, and "switch" state to replace the current state.

Modules

level
The Game Level State handles loading, reloading, and progressing levels.

loading
Because it'd be awkward if you didn't load something in time

menu
The main menu!

Structs

LevelProgression
LevelProgression tracks what the current level is. Because of the way we have implemented level progression, this is vitally important.

Levels
The "Levels" struct tracks which levels we have loaded.