Module rust_action_heroes::state::loading

Because it'd be awkward if you didn't load something in time

The initial state in the game is the loading state.

In theory it transitions to MenuState when all assets are done loading, but in practice we only load a few assets in the Loading state. More than we used to -- we used to load spritesheets everytime we loaded a level. We're not that bad any more, but we still have some room for improvement.

Structs

LoadingState
Handles loading level files and the sprite sheet.