Module room_booking_service::storage

Modules

room_booking

Statics

BOOKING_LIST 🔒

A lazily initialised HashMap containing the list of bookings held by the system.

SNAPSHOT_PATH 🔒

The path used to store a snapshot of the stored booking data.

Functions

create

Create a new booking.

fetch_all

Fetch a list of all room bookings.

fetch_by_check_in_date

Fetch a list of bookings with a specific check in date.

fetch_by_customer_id

Fetch a list of bookings made by a specific customer.

fetch_by_id

Fetch a booking using a booking id.

fetch_by_room_type_id

Fetch a list of bookings made by a specific customer.

load_snapshot

Loads the snapshot from the path defined by SNAPSHOT_PATH into the BOOKING_LIST HashMap.

save_snapshot 🔒

Saves a snapshot of the BOOKING_LIST HashMap to the path defined by SNAPSHOT_PATH. Data is converted to binary for improved storage efficiency.

snapshot_exists

Checks whether a storage snapshot exists in the path defined by SNAPSHOT_PATH.

status

Update the status of a booking.