Trait crossbus::stream::Stream

pub trait Stream<Item>where
    Self: Actor,{
    // Required method
    fn action(&mut self, msg: Item, ctx: &mut Context<Self>);

    // Provided methods
    fn started(&mut self, _ctx: &mut Context<Self>) { ... }
    fn state(&mut self, _ctx: &mut Context<Self>) -> StreamingState { ... }
    fn spawn_stream<S>(&mut self, ctx: &mut Context<Self>, stream: S) -> Handle
       where S: Stream + 'static,
             Self: Stream<S::Item> { ... }
    fn aborted(&mut self, _ctx: &mut Context<Self>) { ... }
    fn paused(&mut self, _ctx: &mut Context<Self>) { ... }
    fn resumed(&mut self, _ctx: &mut Context<Self>) { ... }
    fn finished(&mut self, _ctx: &mut Context<Self>) { ... }
}

An abstraction for Actor’s stream routine

Required Methods

fn action(&mut self, msg: Item, ctx: &mut Context<Self>)

action to do for the Stream Item

Provided Methods

fn started(&mut self, _ctx: &mut Context<Self>)

called before the actor emit the first Strem Item

fn state(&mut self, _ctx: &mut Context<Self>) -> StreamingState

change the state of stream to abort/pause/resume the stream accordingly

Real-Time control or more elaborated execution could be achieved right here

fn spawn_stream<S>(&mut self, ctx: &mut Context<Self>, stream: S) -> Handlewhere S: Stream + 'static, Self: Stream<S::Item>,

spawn stream to the actor

fn aborted(&mut self, _ctx: &mut Context<Self>)

called after the actor aborts the stream

fn paused(&mut self, _ctx: &mut Context<Self>)

called after the actor pause the stream

fn resumed(&mut self, _ctx: &mut Context<Self>)

called after the actor resume the stream

fn finished(&mut self, _ctx: &mut Context<Self>)

called after the actor send the last Strem Item