Struct bpf_bindings::Alu

pub struct Alu {
    // some fields omitted
}

ALU Commands.

Result is always Assigned to A. Each ALU instruction can only perform 1 operation (ADD (+), MUL (*), SUB (-), DIV (/), AND (&), OR (|), LeftShift (<<), RightShift (>>)). The API will not throw an error if the user chains mutiple values, but the kernel will.

A simple example is:

  use bpf_bindings::{Alu, bpf_insn};

  //A = A + K
  let x: bpf_insn = Alu::add().constant(5);

  //A = A | X (bitwise)
  let y: bpf_insn = Alu::or().x_reg();

The API always starts by declaring the operation. Then the developer will state where the operand will come from, be it X or K.

Methods

impl Alu

fn add() -> Alu

Start an ADD Instruction

fn mul() -> Alu

Start a MUL Instruction

fn sub() -> Alu

Start a SUB Instruction

fn div() -> Alu

Start a DIV Instruction

fn and() -> Alu

Start an AND Instruction

fn or() -> Alu

Start an OR Instruction

fn sl() -> Alu

Start a LeftShift Instruction

fn sr() -> Alu

Start a RightShift Instruction

fn not() -> bpf_insn

A = -A The Not Instruction

fn constant(self, k: u32) -> bpf_insn

Do the operation against a constant

fn x_reg(self) -> bpf_insn

Do the operation against reg X (Index Register)