ecdsa_proof

Struct Swiyu

pub struct Swiyu { /* private fields */ }

A simple representation of the Swiyu app. It needs to be set up correctly by the user, which of course usually is done automatically. But for our example we want to see how the public key gets transferred from the SecureElement to the Swiyu app. In the same way you’ll have to add the VerifiedCredential manually.

Implementations

impl Swiyu

pub fn new(setup: PublicSetup) -> Self

Create a new Swiyu app. Of course the real app will do so much more to initialize itself. But here it’s mostly a placeholder to show which actor does which parts.

pub fn add_vc(&mut self, key_id: usize, credential: VerifiedCredential)

Add a VerifiedCredential. In this simple implementation, only one VerifiedCredential can exist.

pub fn bbs_presentation( &mut self, message: VerifierMessage, reveal: Vec<usize>, ) -> BBSPresentation

Returns a BBSPresentation which is a proof that the holder knows the messages signed by the Issuer. In addition to the normal protocol, this adds two PublicCommitments to the proof, which hide the public key of the holder. When creating a BBSPresentation, the PublicCommitments are Open, meaning that their random values are accessible. Before sending a BBSPresentation to an untrusted entity, one must call BBSPresentation::close() to remove the random values.