Module rust_action_heroes::system::switches

How the game knows if all the switches are down

The only thing here is the SwitchSystem struct, so go read about that!

Structs

SwitchSystem
SwitchSystem Writes out the SwitchEvent::AllSwitchesDown event when all switches are covered up; i.e., all switches are "down".

Enums

SwitchEvent
We have this SwitchEven struct in this file for no good reason. I really should move it...