Module liva_parser::tokens

Collection of all parsers and structs which represent tokens, operators and keywords

Structs

BINOP_PRECEDENCE

Vector of binary operators which show the precedence of the operators with which binary expressions should be evaluated

UNOPS

Vector of unary operators which show the precedence of the operators with which unary expressions should be evaluated

Enums

Operator

All Operators which are used in the language for binary expressions

UnOperator

All Operators which are used in the language for unary expressions

Constants

KEYWORDS

Reserved keywords for the liva lang, which can not be used as identifiers

Functions

add

Macro generated function Token parser to parse token

class

Macro generated function Token parser to parse token

comma

Macro generated function Token parser to parse token

div

Macro generated function Token parser to parse token

dot

Macro generated function Token parser to parse token

end

Macro generated function Token parser to parse token

equal

Macro generated function Token parser to parse token

external

Macro generated function Token parser to parse token

fun

Macro generated function Token parser to parse token

greater_eq_than

Macro generated function Token parser to parse token

greater_than

Macro generated function Token parser to parse token

import

Macro generated function Token parser to parse token

land

Macro generated function Token parser to parse token

las

Macro generated function Token parser to parse token

ldo

Macro generated function Token parser to parse token

left_bracket

Macro generated function Token parser to parse token

left_paren

Macro generated function Token parser to parse token

lelse

Macro generated function Token parser to parse token

less_eq_than

Macro generated function Token parser to parse token

less_than

Macro generated function Token parser to parse token

lfor

Macro generated function Token parser to parse token

lif

Macro generated function Token parser to parse token

lin

Macro generated function Token parser to parse token

llet

Macro generated function Token parser to parse token

lor

Macro generated function Token parser to parse token

lreturn

Macro generated function Token parser to parse token

lwhile

Macro generated function Token parser to parse token

mul

Macro generated function Token parser to parse token

right_bracket

Macro generated function Token parser to parse token

right_paren

Macro generated function Token parser to parse token

sub

Macro generated function Token parser to parse token

unequal

Macro generated function Token parser to parse token