Trait reactor_rt::ReactorInitializer

pub trait ReactorInitializer: ReactorBehavior {
    type Wrapped;
    type Params;

    const MAX_REACTION_ID: LocalReactionId;

    fn assemble(
        args: Self::Params, 
        assembler: &mut AssemblyCtx<'_>
    ) -> Result<Self, AssemblyError>
    where
        Self: Sized;
}

Wrapper around the user struct for safe dispatch.

Fields are

the user struct,

ctor parameters of the reactor, and

every logical action and port declared by the reactor.

Associated Types

type Wrapped

Type of the user struct

type Params

Type of the construction parameters

Associated Constants

const MAX_REACTION_ID: LocalReactionId

Exclusive maximum value of the local_rid parameter of [ErasedReactorDispatcher.react_erased].

Required methods

fn assemble(
args: Self::Params,
assembler: &mut AssemblyCtx<'_>
) -> Result<Self, AssemblyError> where
Self: Sized,

Assemble the user reactor, ie produce components with uninitialized dependencies & make state variables assume their default values, or else, a value taken from the params.