Module rustracer::transformation

3D Homogeneous Transformation module.

Provides Matrix and Transformation struct.

Structs

Matrix 🔒

4D Matrix struct.

Transformation

3D Homogeneous Transformation struct.

Constants

IDENTITY_MATRIX 🔒

4D Identity matrix.

Functions

rotation_x

Return a Transformation object encoding a rotation around the X axis.

rotation_y

Return a Transformation object encoding a rotation around the Y axis.

rotation_z

Return a Transformation object encoding a rotation around the Z axis.

scaling

Return a Transformation object encoding a scaling.

translation

Return a Transformation object encoding a rigid translation.