Struct grace_lib::type_checking::context::Context

pub struct Context {
    pub root_id: usize,
    pub scopes: HashMap<usize, Scope>,
    pub containing_scopes: HashMap<usize, usize>,
    pub type_map: HashMap<usize, Type>,
    pub defined_types: HashMap<Identifier, Type>,
    pub gradual_constraints: HashMap<usize, Vec<Type>>,
    pub traits: HashMap<Identifier, Trait>,
    pub trait_implementations: HashMap<(Identifier, Type), HashMap<Identifier, Type>>,
}

The full scoping and typing context for a compilation.

Fields
root_id: usize

The ID of the root scope of the context.

scopes: HashMap<usize, Scope>

A map from Scope IDs to Scopes.

containing_scopes: HashMap<usize, usize>

A map from Node IDs to Scope IDs. Each node that modifies scope maps to the scope it’s contained in.

type_map: HashMap<usize, Type>

A map from Node IDs to types.

defined_types: HashMap<Identifier, Type>

The user-defined types

gradual_constraints: HashMap<usize, Vec<Type>>

A vector containing all the gradual types in context.

traits: HashMap<Identifier, Trait>

A vector of all the traits in context.

trait_implementations: HashMap<(Identifier, Type), HashMap<Identifier, Type>>

The set of pairs (trait_name, type) available, where type implements trait_name (or the trait corresponding to trait_name, rather). The value is the map from method implementations to types.

Implementations

impl Context

Constructors

pub fn builtin() -> Context

Create a context containing all builtin types and functions.

pub fn empty() -> Context

Create a context that contains only an empty scope.

pub fn new_context(scope: Scope, type_map: HashMap<usize, Type>) -> Context

Create a new Context.

impl Context

Basic methods

pub fn add_type(&mut self, id: usize, t: Type)

Record the type of a node.

pub fn define_type(&mut self, name: Identifier, t: Type)

Define a named type.

pub fn get_defined_type(&self, name: &Identifier) -> Result<Type, GraceError>

Get the type associated with a particular name.

pub fn get_type( &self, scope_id: usize, name: &Identifier ) -> Result<Type, GraceError>

Get the type of the identifier in the given scope.

pub fn get_node_type(&self, node_id: usize) -> Result<Type, GraceError>

Get the type of the given node.

impl Context

Scoping methods

pub fn get_scope(&self, scope_id: usize) -> Result<&Scope, GraceError>

Access a scope.

pub fn get_mut_scope( &mut self, scope_id: usize ) -> Result<&mut Scope, GraceError>

Get a mutable reference to a Scope.

pub fn new_scope(&mut self, scope: Scope) -> usize

Create a new scope, returning the ID.

pub fn append_declaration( &mut self, scope_id: usize, name: &Identifier, stmt: &Node<Stmt> )

Add a statement to a scope.

pub fn append_import(&mut self, import: &Import)

Add an import to a scope.

pub fn get_declaring_scope( &self, scope_id: usize, name: &Identifier ) -> Result<usize, GraceError>

Get the scope that declares the given identifier.

pub fn get_declaration( &self, scope_id: usize, name: &Identifier ) -> Result<&CanModifyScope, GraceError>

Get the declaration of a name within a scope.

pub fn get_struct_and_trait( &self, scope_id: usize ) -> (Option<Identifier>, Option<Identifier>)

impl Context

Typechecking

pub fn resolve_attribute( &self, base_type: &Type, name: &Identifier ) -> Result<Type, GraceError>

Resolve an attribute access within the current context.

pub fn resolve_self_type( &self, base_type: &Type, scope_id: usize ) -> Result<Type, GraceError>

Modify a Type::Self so it contains whatever Self actually is.

pub fn trait_information( &self, base_type: &Type, name: &Identifier ) -> Option<Identifier>

Check if this is a trait method call or an attribute access.

pub fn check_trait_method_call( &mut self, scope_id: usize, trait_name: &Identifier, method_name: &Identifier, implementing_type: &Type, arg_types: Vec<&Type> ) -> Result<Type, GraceError>

Check that a trait method call is valid, and get the return type.

pub fn update_gradual(&mut self, gradual_id: usize, constraint: &Type) -> bool

Update a gradual type with a new constraint.

pub fn check_function_types( &self, expected_type: &Type, func_type: &Type ) -> Result<(), GraceError>

pub fn _type_matches( &mut self, scope_id: usize, expr_t: &Type, desired_type: &Type ) -> bool

pub fn check_grad_and_ref_equality( &mut self, scope_id: usize, expr_t: &Type, desired_type: &Type ) -> Result<(), GraceError>

Check if the type of expr is a subtype of desired_type. For types that do not include any gradual or refinement types this is equivalent to equality. For refinement types we have the additional requirement that the refinement constraints be satisfied. For gradual types at least one of the possible gradual types must be a subtype of the desired type.

pub fn merge(&self, t1: &Type, t2: &Type) -> Result<Type, GraceError>

Merge two types if they’re compatible.

impl Context

Helpers

pub fn all_variable_names(&self) -> Vec<Identifier>

Get every variable name in the scope

pub fn all_names_and_types(&self) -> Vec<(Identifier, Result<Type, GraceError>)>

Get every variable name and its type.

pub fn print_identifier_map(&self)

Print every name and its type

pub fn print_implemented_traits(&self)

Print every trait implementation