Module accumulator::util

Miscellaneous functions used throughout the library.

Traits

TypeRep
Pseudo-type-level programming. This trait allows us to reflect "type-level" (i.e. static) information at runtime.

Functions

divide_and_conquer
Folds over xs but in a divide-and-conquer fashion: Instead of F(F(F(F(acc, a), b), c), d)) this computes F(acc, F(F(a, b), F(c, d))).

int
Convenience wrapper for creating Rug integers.

prime_hash_product
Hashes its arguments to primes and takes their product.

shamir_trick
Computes the (xy)th root of g given the xth and yth roots of g and (x, y) coprime.

solve_linear_congruence
Solves a linear congruence of form ax = b mod m for the set of solutions x. Solution sets are characterized by integers mu and v s.t. x = mu + vn and n is any integer.