Module matrs::vec

A vector is of n elements, and is stored in a contiguous block of memory.

Supported operations:

Indexing

Slicing

Iteration

Addition

Subtraction

Multiplication

Division by scalar

Dot product

Cross product

Example

use matrs::vec::Vec;
let mut v = Vec::<f32, 3>::new();
v.set(0, 1.0);
v.set(1, 2.0);
v.set(2, 3.0);
assert_eq!(*v.get(0), 1.0);
assert_eq!(*v.get(1), 2.0);
assert_eq!(*v.get(2), 3.0);

Structs

Vec

Type definition for a vector It is a fixed size vector It is used to implement linear algebra for neural nets and similar