Struct rustracer::vector::Vector

pub struct Vector {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

3D Vector struct.

Fields
x: f32

x component

y: f32

y component

z: f32

z component

Implementations

impl Vector

pub fn neg(self) -> Self

Return the reversed vector.

pub fn dot(self, other: Vector) -> f32

Compute the dot product between two vectors.

pub fn squared_norm(self) -> f32

Return the squared norm (Euclidean length) of a vector.

This is faster than norm if you just need the squared norm.

pub fn norm(self) -> f32

Return the norm (Euclidean length) of a vector.

pub fn normalize(self) -> Self

Modify the vector’s norm so that it becomes equal to 1.

And return the normalized vector.