layers
Module view

Helper struct to build complex layer hierarchies that change over time

It is a wrapper around a state S, that can change overtime, and a render function that generates a LayerTree from a given state.

The view keeps track of the last state and only updates the layertree if the state has changed. A View is mounted on a root layer and appends and removes layers based on LayerTree rendered.

The view keeps a cache of the layers that are rendered by the engine to optimise their creation.

Macros

layer_trees

layer_trees_opt

Structs

LayerTree

A struct that represents a definition of a layer hierearchy that can be rendered by the engine into layers key value is used to optimize the rendering of the layer by reusing the layer when the key is the same

LayerTreeBuilder

Builder for LayerTree.

View

Enums

LayerTreeBuilderError

Error type for LayerTreeBuilder

Traits

BuildLayerTree

A trait for structs that accept a layertree. It is implemented for Layer it generates and updates the properties of a hierarchy of layers described by the layertree

RenderLayerTree

A trait for structs that can produce into a layertree

ViewRenderFunction

Type Aliases

PostRenderHook

PreRenderHook

A View<S> is a struct to support the creation of complex hierarchies of layers that can be rendered by the engine.