Crate peroxide

peroxide is comprehensive numerical library for Rust.

Components

peroxide has various components for scientific computation.

Linear Algebra (with BLAS & LAPACK)

Matrix operations

+,-,*,/

LU Decomposition, Determinant, Inverse

QR Decomposition (O3 feature needed)

Singular Value Decomposition (O3 feature needed)

Cholesky Decomposition (O3 feature needed)

Reduced Row Echelon Form

Vector operations

Eigenvalue, Eigenvector algorithms

Statistics

Statistical operations

mean, var, sd

factorial, P, C, H

Confusion Matrix & metrics

Distributions

Bernoulli

Uniform

Binomial

Normal

Gamma

Beta

Student’s t

Special functions (Using puruspe crate)

Gaussian

Gamma

Beta

Error

Incomplete Gamma

Incomplete Beta

Automatic Differentiation

Taylor mode forward AD

Numerical Utils

Interpolation

Spline

Polynomial

Lanczos Approximation

Numerical Integrations

Optimization

Gradient Descent

Levenberg-Marquardt

Root Finding

Bisection

False Position (Regula falsi)

Secant

Newton

Differential Equations

Explicit

Runge-Kutta 4th order

Euler methods

Implicit

Backward Euler

Gauss-Legendre 4th order

Communication with Python

Plot with matplotlib

DataFrame

Read & Write with parquet or netcdf or csv format

Macros

R macros

Matlab macros

Julia macros

And all these things are built on mathematical traits.

Traits

Functional Programming tools

General algorithms

Mathematics

Mutable tools

Number & Real

Pointer

Stable

Quick Start

Cargo.toml

Run below commands in your project directory

Default

cargo add peroxide

OpenBLAS

cargo add peroxide --features O3

Plot

cargo add peroxide --features plot

NetCDF dependency for DataFrame

cargo add peroxide --features nc

CSV dependency for DataFrame

cargo add peroxide --features csv

Parquet dependency for DataFrame

cargo add peroxide --features parquet

All features

cargo add peroxide --features "O3 plot nc csv parquet"

Import all at once

Peroxide has two options.

prelude : To simple use

fuga : To control numerical algorithms

To see differences, follow above two links.

You can import all functions & structures at once

prelude

#[macro_use]
extern crate peroxide;
use peroxide::prelude::*;

fn main() {
    // Write what you want
}

fuga

#[macro_use]
extern crate peroxide;
use peroxide::fuga::*;

fn main() {
    // Write what you want
}

Useful tips for features

After 0.28.0, dataframe feature is replaced by nc feature.

If you want to use QR or SVD or Cholesky Decomposition then should use O3 feature (there are no implementations for these decompositions in default)

If you want to write your numerical results, then use parquet or nc features (corresponding to parquet or netcdf format. (It is much more effective than csv and json.)

After 0.23.0, there are two options - fuga, prelude. Choose proper option for your computations.

To plot, use parquet or nc feature to export data as parquet or netcdf format and use python to draw plot.

plot feature has limited plot abilities.

To read parquet file in python, use pandas & pyarrow libraries.

There is a template of python code for netcdf. - Socialst

Modules

fuga

Choose what you want.

macros

Useful macros

ml

Machine learning tools

numerical

Differential equations & Numerical Analysis tools

prelude

Do not disturbed. Just use.

special

Special function module

statistics

Statistical Modules

structure

Main structures for peroxide

traits

util

Utility - plot, print, pickle and etc.

Macros

c

R like concatenate (Type: Vec<f64>)

cbind

R like cbind

dnorm

R like dnorm

dt

R like dt

eye

MATLAB like eye - identity matrix

hstack

linspace

MATLAB like linspace

lm

R like lm

matrix

More R like Matrix constructor (Macro)

pnorm

R like pnorm

pt

R like pt

rand

MATLAB like rand - random matrix

rbind

R like rbind

rnorm

R like random normal

rt

R like random Student’s t

runif

R like random uniform

seq

R like seq macro

vstack

zeros

MATLAB like zeros - zero matrix