Crate test_data_generation

Modules

configs
Configs

data

data_sample_parser
The data_sample_parser module provides functionality to read sample data, parse and analyze it, so that test data can be generated based on profiles.

macros

params

profile

Macros

random_percentage