Crate feroxide

Modules

data_atoms

data_ions

data_molecules

data_sep

display_impls

Macros

ion_from_atom

ion_from_molecule

is_letter

is_lower

is_number

is_upper

is_vowel

is_whitespace

molecule_from_atom

to_number

Structs

Atom
An atom

Container
A container for elements

ContainerCompound
A compound for containers

ELECTRON

ElemReaction
An elementair reaction (not containing ions)

Ion
An Ion

Molecule
A molecule

MoleculeCompound
A compound of a molecule

ReactionCompound
A reaction compound

ReactionSide
A side of a reaction

RedoxReaction
A Redox reaction

Traits

Element

Properties

Reaction

Functions

charge_of_atom
Get the charge an atom has based on its group

gcd
Calculate Greatest Common Divisor (GCD), using Euclides' algorithm

ion_superscript
Convert a number to ionic superscript notation The difference with normal superscript notation is that the 1 is omitted, also, ionic superscript supports negative numbers (of which the sign is put at the end, instead of at the beginning)

number_to_greek
Convert a number to greek notation

number_to_roman
Convert a number to roman notaion

rc_to_cc
Convert a given ReactionCompound into a ContainerCompound

subscript
Convert a number to subscript notation

superscript
Convert a number to superscript notation See also ion_superscript

Type Definitions

AtomGroup
should fit 1 to 18

AtomMass
sould fit 0.0 to 294.0+, with normal precision [ gram / mol ]

AtomNumber
should fit 0 to 118+

Energy
should fit 0.0 to 1e5+, with high precision [ Joule ]

IonCharge
should fit -7 to 7

Mass
should fit 0.0 to 1e5+, with high precision [ gram ]

Moles
should fit 0.0 to 1e5+, with high precision [ mol ]

SEP
should fit -5.0 to 5.0, with normal precision [ volt ]