Crate mlx_sys

Modules

array

TODO: add bindings to constructors that takes Primitives?

backend

device

dtype

ffi

fft

linalg

macros

ops

random

stream

transforms

The function pointer used in functions like grad() are used to eventually compute the output array which is then used to compute the gradient. So theoretically we could instead pass a rust function that is callable from C++.

types

utils

Macros

cxx_vec

Enums

Optional

StreamOrDevice