Crate docs

Fast and easy use docker to create a container

There are multiple ways to install Docker:

MAC user: Download Docker from the docker hub

WINDOWS user: First install wsl, then download and install Docker Desktop.

For windows user

wsl --install

Afterwards, double-click the Docker Desktop installer to complete the installation.

Docker start

Once Docker is installed, you can verify the installation by running:

docker --version

And test it with:

docker run hello-world

If everything is set up correctly, you will see a welcome message. You are now ready for the next steps.

Docker container

Let’s begin by pulling a simple container:

docker pull ubuntu

Then run the container interactively:

docker run -it ubuntu bash

you can check all running container by using the following command

docker ps

if you want to put the container into image you can use the following command

docker commit container_id image_name

Replace <container_id> with the container’s ID (from docker ps) and <image_name> with your desired image name.

Exporting an Image to a TAR File

If you want to share an image or store it as a file, you can export it (save) as a tar file:

docker save -o <path_to_file>.tar <image_name>