Crate rustwtxt

Functions

get_statuses
Pull the individual statuses from a remote twtxt.txt file into a std::collections::BTreeMap, with the dates parsed into a chrono::DateTime<FixedOffset>

parse_metadata
This parses out the information in the "== Metadata ==" section of a given twtxt.txt file. You'll have to play with the trim offset here. Often there will be unprintable characters, such as newlines, appended despite the output being trim()-ed before return.

pull_twtxt
Pulls the actual twtxt.txt file from the specified URL.

Type Definitions

Result