Crate bpf_bindings

Simple Bindings to BPF

The Berkeley Packet Filter is a simple 3 register RISC-like Virtual Machine. There are 3 registers. PC (Program Counter) Think of this like an instruction pointer. A (Accumulator) this is a general purpose register. X (index) this is a general purpose register. All registers are 32bits wide. There are two general purpose memory locations. P[] (The Packet) immutable read only and M[] (Scratch Pad) read/write memory for the user to manipulate. This memory is indexed like a byte array. The general notation I will use is P[INDEX:READ_WIDTH], this is similiar to the python slice notation. For example A = P[15:2] will store 2 bytes (Index 15, and 16) of the packet in the lower two bytes of The Accumulator Register.

For a more indepth explaination see: https://www.freebsd.org/cgi/man.cgi?bpf(4)

Structs

Alu
ALU Commands.

CopyReg
Copy Reg

Jmp
Jump Instructions

LoadA
Load Accumulator Register

LoadX
Load Index Register

Ret
Exit Program

StoreReg
Store Register

bpf_factory
Berkeley Packet Filter Program Factory

bpf_insn
Berkeley Packet Filter Instruction

bpf_program
Berkeley Packet Filter Program

Enums

SliceSize
This enum is used to set the size of a read or write. Word reads 32bits, 4 bytes. HalfWord reads 16bits, 2 bytes. Byte sets... well 1 byte.