Crate radar_server

Radar Server

The one binary to rule them all for all things related to Professor John Mower’s radar system.

Goals:

When running as a Server:

Source radar data from the radar hardware in such a way that adding support for new radar systems is easy.

Source simulated radar data from a random function.

Send radar data to clients through a ZMQ Publisher Subscriber connection.

Listen as a ZMQ router, and adjust radar control commands from authorized clients.

When running as an Archiver:

Subscribe to a Server’s published data.

Archive that data in an HDF5 file.

Send live data to the appropriate clients as a ZMQ router.

Listen as a ZMQ router for requests for archived data.

Respond as a ZMQ router with archived data.

Listen as a ZMQ router and forward server commands as a dealer.

When running as a Transformer:

Receive live or archived data from the Archiver as a ZMQ dealer.

Listen as a ZMQ router and forward server commands as a dealer.

Listen as a ZMQ router client requests.

Request the appropriate live or historical data from the archiver.

Transform the data if requested.

Send the potentially transformed, live or historical data to the client as a router.

Current ToDos

Check if the Subscriber trait implementation actually needs to be a loop. (It probably doesn’t)

Missing ZMQ Router Implementation.

Send live packets as Archiver. (Blocked by Missing ZMQ Router Implementation)

Archiver is not Multithreaded.

Send Archived packets as Archiver. (Blocked by Missing ZMQ Router Implementation and Archiver is not Multithreaded)

Figure out what the transformer actually does.

Implement Transform mode. (Blocked by Figure out what the transformer actually does)

Add NUMSAMPLES to the network spec

Move NUMSAMPLES in source_data into one of the radar_packet structs. (Blocked by Add NUMSAMPLES to the network spec)

Acquire hardware/hardware simulation

Source Real Server data. (Blocked by Acquire hardware/hardware simulation)

Finalize Radar Blanking Values. (Blocked by Acquire hardware/hardware simulation)

Implement Radar Control. (Blocked by Finalize Radar Blanking Values)

Finalize radar control security. (Blocked by Implement Radar Control)

Add cryptography data to radar_packet (Blocked by Finalize radar control security)

Implement radar control (Blocked by Add cryptography data to radar packet)

Implement Server mode. (Blocked by Sourcing Real Server Data and Implement radar control)

Modules

plugins 🔒

The plugins module contains all logic and datastructure submodules

Statics

CONTROL_ADDRESS 🔒

The UNUSED address to which all control data is sent and received, tcp://*:5555

DATA_ADDRESS 🔒

The address to which all radar data is sent, tcp://*:5556

SUBSCRIPTION_ADDRESS 🔒

The address to which all archiver data is received, tcp://localhost:5556

TRANSFORMER_ADDRESS 🔒

The UNUSED address on which the archiver and transformer communicate, tcp://localhost:5557

Functions

main 🔒

Main handles argument parsing and calling the necessary submodules.