Trait citadel_sdk::backend_kv_store::BackendHandler

pub trait BackendHandler: TargetLockedRemote {
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn remove<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn set<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 str,
        value: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn get_all<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<HashMap<String, Vec<u8>>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn remove_all<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<HashMap<String, Vec<u8>>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

Contains a trait for persisting application-level data in a K,V store that is unique for this particular connection

Provided Methods

fn get<'life0, 'life1, 'async_trait>(
&'life0 mut self,
key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

Gets a value from the backend

fn remove<'life0, 'life1, 'async_trait>(
&'life0 mut self,
key: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

Removes a value from the backend, returning the previous value

fn set<'life0, 'life1, 'async_trait>(
&'life0 mut self,
key: &'life1 str,
value: Vec<u8>
) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Send + 'async_trait,

Stores a value in the backend, either creating or overwriting any pre-existing value

fn get_all<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<HashMap<String, Vec<u8>>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Obtains the K,V map for this application

fn remove_all<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<HashMap<String, Vec<u8>>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Obtains a list of K,V pairs such that needle is a subset of the K value