Trait rustracer::camera::FireRay

pub trait FireRay {
    // Required method
    fn fire_ray(&self, u: f32, v: f32) -> Ray;
}

Trait for fire a Ray through a camera.

You should redefine it in derived classes.
Fire a ray that goes through the screen at the position (u, v).
The exact meaning of these coordinates depend on the projection used by the camera.

Required Methods

fn fire_ray(&self, u: f32, v: f32) -> Ray