Module citadel_sdk::prefabs::client::peer_connection

A kernel that assists in allowing multiple possible peer-to-peer connections

Structs

AddedPeer

PeerConnectionKernel

A kernel that connects with the given credentials. If the credentials are not yet registered, then the Self::new_register function may be used, which will register the account before connecting. This kernel will only allow outbound communication for the provided account

PeerConnectionSetupAggregator

Allows easy aggregation of UserIdentifier’s and custom settings for the connection request