citadel_sdk::prefabs::client
Module single_connection

A kernel that only makes a single client-to-server connection

Structs

SingleClientServerConnectionKernel

This SingleClientServerConnectionKernel is the base kernel type for other built-in implementations of NetKernel. It establishes connections to a central node for purposes of NAT traversal and peer discovery, and depending on the application layer, can leverage the client to server connection for other purposes that require communication between the two.