Module citadel_sdk::prelude

Convenience import for building applications

Re-exports

pub use crate::backend_kv_store::BackendHandler;

pub use crate::prefabs::client::PrefabFunctions;

pub use crate::responses;

pub use crate::remote_ext::*;

Structs

AccountManager

The default manager for handling the list of users stored locally. It also allows for user creation, and is used especially for when creating a new user via the registration service.

ArgonDefaultServerSettings

ConnectFail

ConnectSuccess

ConnectToHypernode

DeRegistration

DeregisterFromHypernode

Disconnect

DisconnectFromHypernode

ExplicitPanicFuture

Ensures that if a panic occurs in a task, the panic backtrace prints and halts the program

FcmKeys

GroupBroadcastCommand

GroupChannel

GroupChannelCreated

GroupChannelRecvHalf

GroupChannelSendHalf

GroupEvent

HdpServer

IfEq

InternalServerError

KernelExecutor

Creates a KernelExecutor

KernelExecutorSettings

Used for fine-tuning parameters within the [KernelExecutor]

MailboxDelivery

MessageGroupKey

MessageGroupOptions

Options for creating message groups

NodeBuilder

Used to construct a running client/peer or server instance

NodeFuture

An awaitable future whose return value propagates any internal protocol or kernel-level errors

NodeRemote

allows convenient communication with the server

ObjectTransferHandle

ObjectTransferHandler

Used to keep track of file transfer progress for either sender or receiver orientation

OutboundRequestRejected

OutboundUdpSender

PeerChannel

PeerChannelCreated

PeerChannelRecvHalf

A stream interface for receiving secure packets NOTE: on drop, if this is used for a P2P connection, disconnection will occur

PeerChannelSendHalf

PeerCommand

PeerEvent

ReKey

ReKeyResult

RegisterFailure

RegisterOkay

RegisterToHypernode

RtdbConfig

For use in the TOML config as well as for during the Post-login phase

SecBuffer

A memory-secure wrapper for shipping around Bytes

SecString

Allows mutable access

SecureProtocolPacket

A thin wrapper used for convenient creation of zero-copy outgoing buffers

SendObject

ServerMiscSettings

Miscellaneous settings for a node serving connections

ServicesConfig

ServicesObject

Passed to the services handler post-login at the server. Intended to be passed to the client afterwards

SessionList

SessionSecuritySettings

SessionSecuritySettingsBuilder

SymmetricIdentifierHandle

A convenience structure for executing commands that depend on a specific registered user

SymmetricIdentifierHandleRef

A reference to a user identifier

Ticket

A type sent through the server when a request is made

UdpChannel

VirtualObjectMetadata

WebRTCCompatChannel

Enums

BackendType

Used when constructing the account manager

ConnectMode

If force_login is true, the protocol will disconnect any previously existent sessions in the session manager attributed to the account logging-in (so long as login succeeds) The default is a Standard login that will with force_login set to false

ConnectProtocol

For saving the state of client-side connections

EncryptionAlgorithm

GenericNetworkStream

GroupBroadcast

GroupBroadcastPayload

GroupType

HypernodeConnectionType

KemAlgorithm

MemberState

NetworkError

The basic error type for this crate

NodeBuilderError

Returned when an error occurs while building the node

NodeRequest

These are sent down the stack into the server. Most of the requests expect a ticket ID in order for processes sitting above the [Kernel] to know how the request went

NodeResult

This type is for relaying results between the lower-level server and the higher-level kernel TODO: Convert to enum structs

NodeType

Used for determining the proper action when loading the server

ObjectTransferOrientation

ObjectTransferStatus

PeerConnectionType

PeerResponse

PeerSignal

ProposedCredentials

When creating credentials, this is required

ReKeyReturnType

SecrecyMode

ServerUnderlyingProtocol

SigAlgorithm

UdpMode

UserIdentifier

A convenience wrapper for passing arguments to functions that require searches for a user

Traits

AlgorithmsExt

IfEqConditional

IfTrueConditional

NetKernel

The NetKernel is the thread-safe interface between the single-threaded OR multi-threaded async protocol and your network application

PairMap

Remote

RuntimeFuture

SyncIO

Conveniant serialization methods for types that #[derive(Serialize, Deserialize)]

TargetLockedRemote

Then

Functions

calculate_optimal_argon_params

Uses: https://www.ory.sh/choose-recommended-argon2-parameters-password-hashing/ “To reach the desired execution time, you can tweak two variables. It is recommended to start with the highest amount of memory possible and one iteration. Reduce the memory until one hashing operation takes less than your desired duration. Next, advance the number of iterations to approach the desired execution time as close as possible”

Type Definitions

VirtualTargetType

For readability

Attribute Macros

async_trait