Struct rustracer::ray::Ray

pub struct Ray {
    pub origin: Point,
    pub dir: Vector,
    pub tmin: f32,
    pub tmax: f32,
    pub depth: u32,
}

Light Ray struct.

Fields
origin: Point

Origin Point of the light ray.

dir: Vector

Vector propagation direction.

tmin: f32

Minimum time of propagation.

tmax: f32

Maximum time of propagation.

depth: u32

Counts the number of reflection occurred.
If a ray is produced by a reflection, its depth is increased by 1 with respect to the original ray.

Implementations

impl Ray

pub fn at(self, t: f32) -> Point

Return the position of the Ray at time t. The time is measured in units of dir vector’s length.