Trait citadel_sdk::remote_ext::ProtocolRemoteTargetExt

pub trait ProtocolRemoteTargetExt: TargetLockedRemote {
    fn send_file_with_custom_chunking<'life0, 'async_trait, T>(
        &'life0 mut self,
        path: T,
        chunk_size: usize
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        T: 'async_trait + Into<PathBuf> + Send,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn send_file<'life0, 'async_trait, T>(
        &'life0 mut self,
        path: T
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        T: 'async_trait + Into<PathBuf> + Send,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn connect_to_peer_custom<'life0, 'async_trait>(
        &'life0 mut self,
        session_security_settings: SessionSecuritySettings,
        udp_mode: UdpMode
    ) -> Pin<Box<dyn Future<Output = Result<PeerConnectSuccess, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn connect_to_peer<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<PeerConnectSuccess, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn register_to_peer<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<PeerRegisterStatus, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn deregister<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn create_group<'life0, 'async_trait>(
        &'life0 mut self,
        initial_users_to_invite: Option<Vec<UserIdentifier>>
    ) -> Pin<Box<dyn Future<Output = Result<GroupChannel, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn list_owned_groups<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<MessageGroupKey>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn rekey<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Option<u32>, NetworkError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

Some functions require that a target exists

Provided Methods

fn send_file_with_custom_chunking<'life0, 'async_trait, T>(
&'life0 mut self,
path: T,
chunk_size: usize
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
T: 'async_trait + Into<PathBuf> + Send,
'life0: 'async_trait,
Self: Send + 'async_trait,

Sends a file with a custom size. The smaller the chunks, the higher the degree of scrambling, but the higher the performance cost. A chunk size of zero will use the default

fn send_file<'life0, 'async_trait, T>(
&'life0 mut self,
path: T
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
T: 'async_trait + Into<PathBuf> + Send,
'life0: 'async_trait,
Self: Send + 'async_trait,

Sends a file to the provided target using the default chunking size

fn connect_to_peer_custom<'life0, 'async_trait>(
&'life0 mut self,
session_security_settings: SessionSecuritySettings,
udp_mode: UdpMode
) -> Pin<Box<dyn Future<Output = Result<PeerConnectSuccess, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Connects to the peer with custom settings

fn connect_to_peer<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<PeerConnectSuccess, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Connects to the target peer with default settings

fn register_to_peer<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<PeerRegisterStatus, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Posts a registration request to a peer

fn deregister<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Deregisters the currently locked target. If the target is a client to server connection, deregisters from the server. If the target is a p2p connection, deregisters the p2p

fn create_group<'life0, 'async_trait>(
&'life0 mut self,
initial_users_to_invite: Option<Vec<UserIdentifier>>
) -> Pin<Box<dyn Future<Output = Result<GroupChannel, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

fn list_owned_groups<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<MessageGroupKey>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Lists all groups that which the current peer owns

fn rekey<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Option<u32>, NetworkError>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Send + 'async_trait,

Begins a re-key, updating the container in the process. Returns the new key matrix version. Does not return the new key version if the rekey fails, or, if a current rekey is already executing