Trait rustracer::render::Solve

pub trait Solve {
    fn solve(&self, ray: Ray, pcg: &mut Pcg) -> Color;
}

A trait for solving rendering equation.

Must accept a Ray and a Pcg, and must return a Color instance telling the color to assign to a pixel in the image.

Note: Pcg parameter will be used only with Renderer::PathTracer.

Required Methods

fn solve(&self, ray: Ray, pcg: &mut Pcg) -> Color