Function docs::routes::about::routes::get_chains_about

pub async fn get_chains_about(
    context: Context<'_>, 
    chain_id: String
) -> Result<Json<String>, ApiError>

/v1/chains/<chain_id>/about
Returns ChainAbout

Chain’s About

The about endpoint provides information of the environmental variables set for the instance of safe-client-gateway. This would allow to identify on which commit and version the last deployment happened and to which safe transaction handlers backend environment the current instance of the gateway is pointing to.

Path

/v1/chains/<chain_id>/about

Query parameters

There are no query parameters for this endpoint