Function docs::routes::transactions::routes::get_transactions_queued

pub async fn get_transactions_queued(
    context: Context<'_>, 
    chain_id: String, 
    safe_address: String, 
    cursor: Option<String>, 
    timezone_offset: Option<String>, 
    trusted: Option<bool>
) -> Result<Json<String>, ApiError>

/v1/chains/<chain_id>/safes/<safe_address>/transactions/queued?<cursor>&<timezone_offset>&<trusted>
Returns a Page of TransactionListItem

Transactions Queued

This endpoint returns all the transactions that are still awaiting execution for a given safe. The list will contain a Next marker if there is a transaction matching the nonce of the safe, which means, that it will be the next transaction to be executed, provided there aren’t any other transactions proposed utilizing the same nonce. If that were, the case a ConflictHeader will be introduced for which the nonce field will hold the conflicting value.

Additionally to the Next marker, there is also Queued. Under this marker, transactions that have a nonce greater than that of the safe are listed. Analogously to the Next section of the list, a ConflictHeader will be introduced for any group of transactions sharing the same nonce.

The structure of the transaction object corresponds to that of a [crate::models::handlers::transactions::summary::TransactionSummary]

A TransactionListItem can be either a Label (containing either Next or Queued), ConflictHeader (with the conflicting nonce) and a Transaction, for which there is a TransactionSummary and a ConflictType associated. The conflict type can have HasNext or End value. These values signal to which extent a group of conflicting transactions spans, ending as soon as a Transaction type item contains a ConflictType::End.

Path

GET /v1/chains/<chain_id>/safes/<safe_address>/transactions/queued?<cursor>&<timezone_offset>&<trusted>

The response is a list of [crate::models::handlers::transactions::summary::TransactionListItem], which is a polymorphic struct. Details follow in the models sections.

Query parameters

<safe_address> should be the checksummed address of the safe to be observed.

<cursor> is the desired page of data to be loaded. Values for this parameter can be either Page.next or Page.previous. WARNING: Don’t fiddle with the values of these 2 fields.

<timezone_offset>: Currently ignored by the gateway.

<trusted>: forwarded directly to the core services. Only for debugging purposes clients should not send it (unless they know what they are doing).