Function docs::routes::transactions::routes::post_transaction

pub async fn post_transaction<'e>(
    context: RequestContext, 
    chain_id: String, 
    safe_address: String, 
    multisig_transaction_request: Result<Json<MultisigTransactionRequest>, Error<'e>>
) -> Result<Json<String>, ApiError>

/v1/chains/<chain_id>/transactions/<safe_address>/propose
returns TransactionDetails

Transaction Proposal

This endpoint provides a way for submitting transactions of any kind in the format expected by the core services. See the example json to see how to submit a cancellation transaction (you would need to supply a nonce, signature and contractTransactionHash appropriate to the transaction you are submitting)

Path

POST /v1/chains/<chain_id>/transactions/<safe_address>/propose

The expected (MultisigTransactionRequest)crate::routes::transactions::models::requests::MultisigTransactionRequest body for this request, can be found in the sections models

Query parameters

No query parameters available for this endpoint.