Function peroxide::special::function::ln_gamma

pub fn ln_gamma(x: f64) -> f64

Logarithm Gamma function

Description

Use Lanczos approximation to implement Gamma function ($g=5, n=7$)

References

Robert Munafo, Coefficients for the Lanczos Approximation to the Gamma Function

Paul Godfrey, A note on the computation of the convergent Lanczos complex Gamma approximation (web page), 2001.